% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{.addCustomLabelsCommands}
\alias{.addCustomLabelsCommands}
\title{Add custom label plotting commands}
\usage{
.addCustomLabelsCommands(x, commands, plot_type)
}
\arguments{
\item{x}{An instance of a \linkS4class{DotPlot} class.}

\item{commands}{A character vector representing the sequence of commands to create the \link{ggplot} object.}

\item{plot_type}{String specifying the type of plot, e.g., \code{"scatter"}, \code{"square"}, \code{"violin"}.}
}
\value{
A character vector containing \code{commands} plus any additional commands required to generate the labels.
}
\description{
Add \link{ggplot} instructions to add custom labels to specified points in a \linkS4class{DotPlot}.
This is a utility function that is intended for use in \code{\link{.generateDotPlot}}.
}
\author{
Kevin Rue-Albrecht, Aaron Lun
}
