% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{aes-utils}
\alias{aes-utils}
\alias{.buildAes}
\title{Generate ggplot aesthetic instructions}
\usage{
.buildAes(
  x = TRUE,
  y = TRUE,
  color = FALSE,
  shape = FALSE,
  size = FALSE,
  fill = FALSE,
  group = FALSE,
  alt = NULL
)
}
\arguments{
\item{x}{A \code{logical} that indicates whether to enable \code{x} in the
aesthetic instructions (default: \code{TRUE}).}

\item{y}{A \code{logical} that indicates whether to enable \code{y} in the
aesthetic instructions (default: \code{TRUE}).}

\item{color}{A \code{logical} that indicates whether to enable
\code{color} in the aesthetic instructions (default: \code{FALSE}).}

\item{shape}{A \code{logical} that indicates whether to enable
\code{shape} in the aesthetic instructions (default: \code{FALSE}).}

\item{size}{A \code{logical} that indicates whether to enable
\code{size} in the aesthetic instructions (default: \code{FALSE}).}

\item{fill}{A \code{logical} that indicates whether to enable
\code{fill} in the aesthetic instructions (default: \code{FALSE}).}

\item{group}{A \code{logical} that indicates whether to enable
\code{group} in the aesthetic instructions (default: \code{FALSE}).}

\item{alt}{Alternative aesthetics, supplied as a named character vector.}
}
\value{
Aesthetic instructions for \code{\link{ggplot}} as a character
value.
}
\description{
Generate ggplot aesthetic instructions
}
\examples{
.buildAes()
}
\seealso{
\code{\link{.scatter_plot}},
\code{\link{.violin_plot}},
\code{\link{.square_plot}}
}
\author{
Kevin Rue-Albrecht
}
