% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotateEnsembl}
\alias{annotateEnsembl}
\title{Title}
\usage{
annotateEnsembl(se, species, keytype, rowdata_col = NULL)
}
\arguments{
\item{se}{}

\item{species}{}

\item{keytype}{}

\item{rowdata_col}{}
}
\description{
Title
}
