% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{availablePanelTypes}
\alias{availablePanelTypes}
\title{Available panel types}
\usage{
availablePanelTypes(row.names = c(NA, "encoded", "decoded"))
}
\arguments{
\item{row.names}{Which column should be set as \code{row.names}.
The default \code{NA} does not set rownames.}
}
\value{
A data.frame.
}
\description{
Displays a data.frame that presents the list of available panel types,
with their decoded and encoded names.
}
\details{
Decoded names are used to define the panels visible at initialization.
Refer to the documentation of the \code{iSEE} function for more information.

Encoded names are used to name the individual HTML elements in the UI.
Those names may be used to anchor tour steps at HTML elements using the \code{rintrojs} package.
}
\examples{
availablePanelTypes()
}
\seealso{
\code{\link{iSEE}}
}
\author{
Kevin Rue-Albrecht
}
