% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_ColDataPlot.R
\docType{methods}
\name{ColDataPlot}
\alias{ColDataPlot}
\alias{ColDataPlot-class}
\alias{.defineParamInterface,ColDataPlot-method}
\alias{.createParamObservers,ColDataPlot-method}
\title{The column data plot panel}
\description{
Plots column data values. What more do I have to say?
}
\section{Constructor}{

\code{ColDataPlot()} creates an instance of a ColDataPlot class.
}

\section{Panel parameters}{

\code{\link{.defineParamInterface}} will create parameter elements for choosing the reduced dimensions to plot.
More details to be added.
}

\author{
Aaron Lun
}
