% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{colDataPlotDefaults}
\alias{colDataPlotDefaults}
\title{Column data plot defaults}
\usage{
colDataPlotDefaults(se, number)
}
\arguments{
\item{se}{A SummarizedExperiment object.}

\item{number}{An integer scalar, specifying the maximum number of column data plots that can be added to the interface.}
}
\value{
A DataFrame containing default settings for parameters of each of \code{number} column data panels.
}
\description{
Create default settings for column data plot panels in the iSEE interface.
}
\details{
Parameters available to column data plots are:
\describe{
\item{\code{YAxis}:}{Character, which column of \code{colData(se)} should be shown on the y-axis?
Defaults to the first entry of \code{colData(se)}.}
\item{\code{XAxis}:}{Character, what type of variable should be shown on the x-axis?
Defaults to \code{"None"}, but can also be \code{"Column data"}.}
\item{\code{XAxisColData}:}{Character, which column of \code{colData(se)} should be shown on the x-axis if \code{XAxis="Column data"}?
Defaults to the first entry of \code{colData(se)}.}
}

All column-based parameters described in \code{?"\link{iSEE point parameters}"} are applicable.
All plot-based parameters described in \code{?"\link{iSEE selection parameters}"} are applicable.
}
\examples{
example(SingleCellExperiment, echo=FALSE) # mock up 'sce'.
colDataPlotDefaults(sce, n=1)
}
\seealso{
\code{?"\link{iSEE point parameters}"},
\code{?"\link{iSEE selection parameters}"}
}
\author{
Aaron Lun
}
