% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{colStatTableDefaults}
\alias{colStatTableDefaults}
\title{Column statistics table defaults}
\usage{
colStatTableDefaults(se, number)
}
\arguments{
\item{se}{A SummarizedExperiment object.}

\item{number}{An integer scalar, specifying the maximum number of column statistics tables that can be added to the interface.}
}
\value{
A DataFrame containing default settings for parameters of each of \code{number} column statistics table panels.
}
\description{
Create default settings for column statistics table panels in the iSEE interface.
}
\details{
Parameters available to col statistics tables are:
\describe{
\item{\code{Selected}:}{Integer, containing the index of the col to be initially selected.
Defaults to the first col, i.e., 1.
Alternatively, a string can be supplied containing the column name.}
\item{\code{Search}:}{Character, containing the initial value of the search field.
Defaults to an empty string.}
\item{\code{SearchColumns}:}{A list containing character vectors of length equal to the number of columns in \code{colData(se)},
specifying the initial value of the search field for each column.
All entries default to an empty string.}
}

All table-based parameters described in \code{?"\link{iSEE selection parameters}"} are applicable.
}
\examples{
example(SingleCellExperiment, echo=FALSE) # mock up 'sce'.
colStatTableDefaults(sce, n=1)
}
\seealso{
\code{?"\link{iSEE selection parameters}"}
}
\author{
Aaron Lun
}
