% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_ColumnDotPlot.R
\docType{methods}
\name{ColumnDotPlot}
\alias{ColumnDotPlot}
\alias{.defineParamInterface,ColumnDotPlot-method}
\alias{.createParamObservers,ColumnDotPlot-method}
\alias{.defineOutputElement,ColumnDotPlot-method}
\alias{.createRenderedOutput,ColumnDotPlot-method}
\title{Column dot plot panel family}
\description{
The column dot plot panel family covers all plot panel types where each column of the \linkS4class{SummarizedExperiment} is represented by a point.
This family is represented by the \code{ColumnDotPlot} virtual class, which has a number of concrete subclasses (e.g., \linkS4class{RedDimPlot}) to direct the construction of specific plots.
We provide a number of useful methods on this virtual class to make it easier for developers to define their own subclasses.
}
\section{Panel parameters}{

\code{\link{.defineParamInterface}} will create parameter elements for visualization and point selection.
More details to be added.
}

\section{Output plot}{

\code{\link{.createOutputElement}} will return the output of \code{\link{plotOutput}} with two-dimensional brushing.
More details to be added.
}

\author{
Aaron Lun
}
