% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observers_select.R
\name{.createMultiSelectionEffectObserver}
\alias{.createMultiSelectionEffectObserver}
\title{Create an observer for multiple selection effect}
\usage{
.createMultiSelectionEffectObserver(
  plot_name,
  by_field,
  type_field,
  saved_field,
  input,
  session,
  pObjects,
  rObjects
)
}
\arguments{
\item{plot_name}{String containing the encoded name of the current panel.}

\item{by_field}{String with the name of the slot containing the name of the transmitting panel.}

\item{type_field}{String with the name of the slot containing the type of selection to receive
(i.e., \code{"Saved"}, \code{"Active"} or \code{"Union"}).}

\item{saved_field}{String with the name of the slot containing the saved history of selections.}

\item{input}{The Shiny input object from the server function.}

\item{session}{The Shiny session object from the server function.}

\item{pObjects}{An environment containing global parameters generated in the \code{\link{iSEE}} app.}

\item{rObjects}{A reactive list of values generated in the \code{\link{iSEE}} app.}
}
\value{
An observer for the input corresponding to \code{"SelectEffect"} is created in the current Shiny session.
A \code{NULL} is invisibly returned.
}
\description{
Create an observer to trigger rerendering upon changes to the (typically visual) effect for multiple selections.
}
\details{
The selection effect is expected to be stored in the \code{SelectionEffect} slot of the current panel.
Changes to the specified effect will trigger rerendering of the current panel if there was a multiple selection transmitted from an upstream panel.
If the specified effect changes to or from \dQuote{Restrict}, all children of the current panel will also be rerendered.

This function is typically called within a \code{\link{.createObservers}} method 
It is exported for use in subclasses as the selection effect is not a property of the \linkS4class{Panel} class (given that not all panels are visual) 
and thus needs to be called separately for each subclass rather than in the parent's method.
}
\seealso{
\code{\link{.createObservers}}, where this function is usually called -
see, for example, the methods for \linkS4class{ColumnDotPlot}, \linkS4class{RowDotPlot} and \linkS4class{ComplexHeatmapPlot}.
}
\author{
Aaron Lun
}
