% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_plots.R
\name{Custom iSEE plots}
\alias{Custom iSEE plots}
\title{Creating custom plots}
\description{
Instructions for creating custom plots for use in \code{\link{iSEE}}.
}
\section{Functions for custom column plots}{

Any function should accept the following arguments:
\itemize{
\item \code{se}, a SingleCellExperiment object.
\item \code{columns}, a character vector of column names, corresponding to a transmitted selection of points.
This will be different from \code{colnames(se)} if the current custom column plot is set to \code{"Restrict"}.
}
In return, it should produce a list containing the following elements:
\describe{
\item{\code{coordinates}:}{A data.frame with one row per entry of \code{columns}.
It should contain the fields \code{X} and \code{Y}, for the x- and y-axis values per row.
Row names must be set to \code{columns}.}
\item{\code{xlab}:}{A string specifying the x-axis label.}
\item{\code{ylab}:}{A string specifying the y-axis label.}
\item{\code{title}:}{A string specifying the plot title.}
}
}

\examples{
# Setting up a function:
CUSTOM <- function(se, rows, columns, ...) {
    kept <- logcounts(se)
    if (!is.null(rows)) {
        kept <- kept[rows,,drop=FALSE]
    }
    if (!is.null(columns)) { 
        kept <- kept[,columns,drop=FALSE]
    }
    out <- prcomp(t(kept), rank.=2)$x
    coords <- data.frame(X=out[,1], Y=out[,2])

    ggplot(coords) + geom_point(aes(x=X, y=Y)) + xlab("PC1") + ylab("PC2")
}

# Setting up the SCE object:
library(scRNAseq)
data(allen)
class(allen)

library(scater)
sce <- as(allen, "SingleCellExperiment")
counts(sce) <- assay(sce, "tophat_counts")
sce <- normalize(sce)
sce <- runPCA(sce)
rowData(sce)$mean_log <- rowMeans(logcounts(sce))
rowData(sce)$var_log <- apply(logcounts(sce), 1, var)

# Setting up the app:
reddim <- redDimPlotDefaults(sce, 1)
rowdat <- rowDataPlotDefaults(sce, 1)
rowdat$XAxis <- "Row data"
rowdat$XAxisRowData <- "mean_log"
rowdat$YAxis <- "var_log"

cdp <- customDataPlotDefaults(sce, 1)
cdp$Function <- "PCA2"
cdp$ColumnSource <- "Reduced dimension plot 1"
cdp$RowSource <- "Row data plot 1"

app <- iSEE(sce, redDimArgs=reddim, rowDataArgs=rowdat, customDataArgs=cdp,
   initialPanels=DataFrame(Name=c("Reduced dimension plot 1", 
       "Row data plot 1", "Custom data plot 1")),
   customDataFun=list(PCA2=CUSTOM))
}
\author{
Aaron Lun
}
