% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{customDataPlotDefaults}
\alias{customDataPlotDefaults}
\title{Custom data plot defaults}
\usage{
customDataPlotDefaults(se, number)
}
\arguments{
\item{se}{A SummarizedExperiment object.}

\item{number}{An integer scalar, specifying the maximum number of custom data plots that can be added to the interface.}
}
\value{
A DataFrame containing default settings for parameters of each of \code{number} custom data plot panels.
}
\description{
Create default settings for custom data plot panels in the iSEE interface.
}
\details{
Data parameters available to custom data plots are:
\describe{
\item{\code{DataBoxOpen}:}{Logical, should the data parameter box be open upon initialization?
Defaults to \code{FALSE}.}
\item{\code{Function}:}{String, which function should be used to generate the ggplot for the curent panel?
Defaults to \code{"---"}, i.e., no coordinates are generated.}
\item{\code{Arguments}:}{String with multiple lines specifying the initial arguments for the function, see \code{vignette("custom", package="iSEE")} for details.
Defaults to an empty string.}
\item{\code{VisibleArgs}:}{String with multiple lines specifying the initial arguments to be shown in the text area.
This can differ from \code{Arguments}, e.g., to list argument names without their values for users to enter.
Defaults to \code{NA}, which means that the value in \code{Arguments} will be shown.}
}

Selection parameters for custom data plots are:
\describe{
\item{\code{SelectBoxOpen}:}{Logical, should the selection parameter box be open upon initialization?
Defaults to \code{FALSE}.}
\item{\code{ColumnSource}:}{Character, which other plot should transmit sample selections to the current plot? 
Defaults to \code{"---"}, which means that no plot is used for point selection.}
\item{\code{RowSource}:}{Character, which other plot should transmit feature selections to the current plot? 
Defaults to \code{"---"}, which means that no plot is used for point selection.}
}
}
\examples{
example(SingleCellExperiment, echo=FALSE) # mock up 'sce'.
customDataPlotDefaults(sce, n=1)
}
\author{
Aaron Lun, Kevin Rue-Albrecht
}
