% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_class.R
\name{dataframe-utils}
\alias{dataframe-utils}
\alias{.findAtomicFields}
\title{Find atomic fields}
\usage{
.findAtomicFields(df)
}
\arguments{
\item{df}{A data.frame or \linkS4class{DataFrame}.}
}
\value{
A character vector of names of atomic fields in \code{df}.
}
\description{
A utility function to find column sin a data.frame or \linkS4class{DataFrame}
that are atomic R types, as most of the app does not know how to handle  more complex types being stored as columns.
An obvious example is in data.frames expected by \code{\link{ggplot}} or \code{\link{datatable}}.
}
\examples{
x <- DataFrame(
    A = rnorm(10),
    B = sample(letters, 10),
    DataFrame = I(DataFrame(
        C = rnorm(10),
        D = sample(letters, 10)
    ))
)

.findAtomicFields(x)
}
\author{
Aaron Lun, Kevin Rue-Albrecht
}
