% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{defaults}
\alias{defaults}
\alias{redDimPlotDefaults}
\alias{geneExprPlotDefaults}
\alias{colDataPlotDefaults}
\title{Plot parameter defaults}
\usage{
redDimPlotDefaults(se, max.plots)

geneExprPlotDefaults(se, max.plots)

colDataPlotDefaults(se, max.plots)
}
\arguments{
\item{se}{A SingleCellExperiment object.}

\item{max.plots}{An integer scalar, specifying the maximum number of 
plots of the corresponding type that can be added to the interface.}
}
\value{
A DataFrame containing default settings for various 
parameters of reduced dimension or gene expression plots.
}
\description{
Create default settings for various plots in the iSEE interface.
}
\section{Reduced dimension plot parameters}{

\describe{
\item{\code{Type}:}{Characater, what entry of \code{reducedDims(se)} should be shown?
By default, the first entry is shown.}
\item{\code{XAxis}:}{Integer, which component should be shown on the x-axis?
Defaults to 1.}
\item{\code{YAxis}:}{Integer, which component should be shown on the y-axis?
Defaults to 2.}
}
}

\section{Gene expression plot parameters}{

\describe{
\item{\code{GeneTable}:}{Character, what gene statistic table should be used to choose a gene to display on the y-axis?
Defaults to an empty string, which means that the first available table will be used.}
\item{\code{Assay}:}{Character, what expression values should be shown on the y-axis?
Defaults to the name of the first assay in \code{se}.}
\item{\code{XAxis}:}{Character, what variable should be shown on the x-axis?
Defaults to \code{"None"}.}
\item{\code{XAxisColData}:}{Character, what column of \code{colData(se)} should be shown on the x-axis if \code{XAxis="Column data"}?
Defaults to the first entry of \code{colData(se)}.}
\item{\code{XAxisGeneExprs}:}{Character, which gene's expression should be shown on the x-axis if \code{XAxis="Gene expression"}? 
Defaults to the name of the first row in \code{se}, using expression values specified in \code{Assay}.}
}
}

\section{Column data plot parameters}{

\describe{
\item{\code{YAxisColData}:}{Character, which column of \code{colData(se)} should be shown on the y-axis?
Defaults to the first entry of \code{colData(se)}.}
\item{\code{XAxis}:}{Character, what variable should be shown on the x-axis?
Defaults to \code{"None"}.}
\item{\code{XAxisColData}:}{Character, which column of \code{colData(se)} should be shown on the x-axis if \code{XAxis="Column data"}?
Defaults to the first entry of \code{colData(se)}.}
}
}

\section{Coloring parameters}{

\describe{
\item{\code{ColorPanelOpen}:}{Logical, should the color parameter panel be open upon initialization?
Defaults to \code{FALSE}.}
\item{\code{ColorBy}:}{Character, what type of data should be used for coloring?
Defaults to \code{"None"}.}
\item{\code{ColorByColData}:}{Character, which column of \code{colData(se)} should be used for colouring if \code{ColorBy="Column data"}? 
Defaults to the first entry of \code{colData(se)}.}
\item{\code{ColorByGeneTable}:}{Character, which gene statistic table should be used to choose a gene to color by, if \code{ColorBy="Gene table"}? 
Defaults to an empty string, which means that the first available table will be used.}
\item{\code{ColorByGeneTableAssay}:}{Character, what expression values should be used for colouring if \code{ColorBy="Gene table"}? 
Defaults to the name of the first assay in \code{se}.}
\item{\code{ColorByGeneText}:}{Character, which gene should be used to choose a gene to color by, if \code{ColorBy="Gene text"}? 
Defaults to an empty string, which means that the first available table will be used.}
\item{\code{ColorByGeneTextAssay}:}{Character, what expression values should be used for colouring if \code{ColorBy="Gene text"}? 
Defaults to the name of the first assay in \code{se}.}
}
}

\section{Brushing parameters}{

\describe{
\item{\code{BrushPanelOpen}:}{Logical, should the brushing parameter panel be open upon initialization?
Defaults to \code{FALSE}.}
\item{\code{BrushOn}:}{Logical, should the plot be transmitting its brush for use in other plots?
Defaults to \code{FALSE}.}
\item{\code{BrushByPlot}:}{Character, which other plot should be used for point selection in the current plot? 
Defaults to an empty string, which means that no plot is used for point selection.}
}
}

\section{Other parameters}{

\describe{
\item{\code{PlotPanelOpen}:}{Logical, should the plot parameter panel be open upon initialization?
Defaults to \code{FALSE}.}
}
}

\examples{
library(scRNAseq)
data(allen)
class(allen)

library(scater)
sce <- as(allen, "SingleCellExperiment")
counts(sce) <- assay(sce, "tophat_counts")
sce <- normalize(sce)
sce <- runPCA(sce)
sce

redDimPlotDefaults(sce, max.plots=5)
geneExprPlotDefaults(sce, max.plots=5)
colDataPlotDefaults(sce, max.plots=5)
}
