% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultTour.R
\name{defaultTour}
\alias{defaultTour}
\title{Define the default tour}
\usage{
defaultTour()
}
\value{
A data.frame where each row is a tour step.
The first column specifies the UI element to be highlighted by the tour,
while the second column contains the tour text.
}
\description{
Define the default tour for the subset of the Allen brain dataset.
This is only available when run on the \code{iSEE(sce)} example in \code{?"\link{iSEE}"}.
}
\examples{
defaultTour()
}
\author{
Aaron Lun
}
