% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{ExperimentColorMap}
\alias{ExperimentColorMap}
\alias{class:ExperimentColorMap}
\alias{ExperimentColorMap-class}
\alias{assayColorMap}
\alias{colDataColorMap}
\alias{rowDataColorMap}
\alias{assayColorMap,ExperimentColorMap,character-method}
\alias{assayColorMap,ExperimentColorMap,numeric-method}
\alias{colDataColorMap,ExperimentColorMap,character-method}
\alias{colDataColorMap,ExperimentColorMap,numeric-method}
\alias{rowDataColorMap,ExperimentColorMap,character-method}
\alias{rowDataColorMap,ExperimentColorMap,numeric-method}
\title{\code{ExperimentColorMap} objects}
\usage{
ExperimentColorMap(assays = list(), colData = list(), rowData = list(),
  all = list(assays = NULL, colData = NULL, rowData = NULL), global = NULL,
  ...)
}
\arguments{
\item{assays}{List of color maps for \code{assays}.}

\item{colData}{List of color maps for \code{colData}.}

\item{rowData}{List of color maps for \code{rowData}.}

\item{all}{List of color maps for \code{all}.}

\item{global}{List of color maps for \code{global}.}

\item{...}{additional arguments passed on to the \code{ExperimentColorMap}
constructor}
}
\description{
\code{ExperimentColorMap} objects
}
\details{
Color maps must all be functions that take at least one argument: the number
of (named) colours to return as a \code{character} vector.
This argument may be ignored in the body of the color map function
to produce constant color maps.
}
\section{Accessors}{


In the following code snippets, \code{x} is an
\code{ExperimentColorMap} object.

\describe{
  \item{\code{assayColorMap(x, i, ...)}:}{Get an \code{assays} colormap.}
  
  \item{\code{colDataColorMap(x, i, ...)}:}{Get a \code{colData} colormap.}
  
  \item{\code{rowDataColorMap(x, i, ...)}:}{Get a \code{rowData} colormap.}
}
}

\examples{

# Example color maps ----

count_colors <- function(n){
  c("black","brown","red","orange","yellow")
}
fpkm_colors <- viridis::inferno
tpm_colors <- viridis::plasma

qc_color_fun <- function(n){
  qc_colors <- c("forestgreen", "firebrick1")
  names(qc_colors) <- c("Y", "N")
  return(qc_colors)
}

# Constructor ----

ecm <- new("ExperimentColorMap",
    assays = list(
        counts = count_colors,
        tophat_counts = count_colors,
        cufflinks_fpkm = fpkm_colors,
        cufflinks_fpkm = fpkm_colors,
        rsem_tpm = tpm_colors
    ),
    colData = list(
        passes_qc_checks_s = qc_color_fun
    )
)

# Accessors ----

assayColorMap(ecm, "undefined") # viridis::viridis(10) [default]
assayColorMap(ecm, "counts") # viridis::plasma(10)
assayColorMap(ecm, "cufflinks_fpkm") # viridis::inferno(10)

colDataColorMap(ecm, "passes_qc_checks_s")
colDataColorMap(ecm, "undefined")

rowDataColorMap(ecm, "undefined")

}
