% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_FeatAssayPlot.R
\docType{methods}
\name{FeatAssayPlot}
\alias{FeatAssayPlot}
\alias{FeatAssayPlot-class}
\alias{.defineParamInterface,FeatAssayPlot-method}
\alias{.createParamObservers,FeatAssayPlot-method}
\title{The feature assay plot panel}
\description{
Plots feature assay values. What more do I have to say?
}
\section{Constructor}{

\code{FeatAssayPlot()} creates an instance of a FeatAssayPlot class.
}

\section{Panel parameters}{

\code{\link{.defineParamInterface}} will create parameter elements for choosing the reduced dimensions to plot.
More details to be added.
}

\author{
Aaron Lun
}
