% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{featAssayPlotDefaults}
\alias{featAssayPlotDefaults}
\title{Feature assay plot defaults}
\usage{
featAssayPlotDefaults(se, number)
}
\arguments{
\item{se}{A SummarizedExperiment object.}

\item{number}{An integer scalar, specifying the maximum number of feature assay plots that can be added to the interface.}
}
\value{
A DataFrame containing default settings for parameters of each of \code{number} feature assay panels.
}
\description{
Create default settings for feature assay plot panels in the iSEE interface.
}
\details{
Parameters available to feature assay plots are:
\describe{
\item{\code{YAxisFeatName}:}{Integer, the index of the feature for which to show the expression on the y-axis if \code{YAxis="Feature name"}.
Defaults to 1, i.e., the first feature in \code{se}.
Alternatively, a string can be supplied containing the name of the feature, i.e., the row name.}
\item{\code{YAxisRowTable}:}{Character, what row statistics table should be used to choose a feature to display on the y-axis?
Any setting will override \code{YAxisFeatName} with the selected row in the chosen table upon initialization of the app.
Defaults to \code{"---"}, which means that no table will be used.}
\item{\code{Assay}:}{Integer, which assay should be used to supply the expression values shown on the y-axis?
Defaults to 1, i.e., the first assay in \code{se}.
Alternatively, a string can also be supplied containing the name of the assay, if \code{assays(se)} has names.}
\item{\code{XAxis}:}{Character, what type of variable should be shown on the x-axis?
Defaults to \code{"None"}, but can also be \code{"Row table"}, \code{"Column data"} or \code{"Feature name"}.}
\item{\code{XAxisColData}:}{Character, what column of \code{colData(se)} should be shown on the x-axis if \code{XAxis="Column data"}?
Defaults to the first entry of \code{colData(se)}.}
\item{\code{XAxisFeatName}:}{Integer, the index of the feature for which to show the expression on the x-axis if \code{XAxis="Feature name"}.
Defaults to 1, i.e., the first feature in \code{se}.
Alternatively, a string can be supplied containing the name of the feature.}
\item{\code{XAxisRowTable}:}{Character, which row statistic table should be used to choose a feature to put on the x-axis if \code{XAxis="Row table"}?
Any setting will override \code{XAxisFeatName} with the selected row in the chosen table upon initialization of the app.
Defaults to \code{"---"}, which means that no table will be used.}
}

All column-based parameters described in \code{?"\link{iSEE point parameters}"} are applicable.
All plot-based parameters described in \code{?"\link{iSEE selection parameters}"} are applicable.
}
\examples{
example(SingleCellExperiment, echo=FALSE) # mock up 'sce'.
featAssayPlotDefaults(sce, n=1)
}
\seealso{
\code{?"\link{iSEE point parameters}"},
\code{?"\link{iSEE selection parameters}"}
}
\author{
Aaron Lun
}
