% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterDTColumn.R
\name{filterDTColumn}
\alias{filterDTColumn}
\title{Filter \pkg{DT} columns}
\usage{
filterDTColumn(x, search)
}
\arguments{
\item{x}{A numeric or character vector, usually representing a column of a data.frame.}

\item{search}{A string specifying the search filter to apply to \code{x}.}
}
\value{
A logical vector indicating which entries are returned.
}
\description{
Filter a vector based on the \pkg{DT} \code{\link{datatable}} widget column search string.
}
\details{
For character \code{x}, \code{search} is treated as a regular expression.

For numeric \code{x}, \code{search} should have the form \code{LOWER ... UPPER}
where all elements in [LOWER, UPPER] are retained.
}
\examples{
filterDTColumn(LETTERS, "A|B|C")

filterDTColumn(runif(20), "0.1 ... 0.5")
}
\seealso{
\code{\link{datatable}} and associated documentation for more details about column searches.
}
\author{
Aaron Lun
}
