% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/utils_name.R
\name{.fullName}
\alias{.fullName}
\alias{.getEncodedName}
\alias{.getFullName}
\title{Get panel names}
\usage{
.fullName(x)

.getEncodedName(x)

.getFullName(x)
}
\arguments{
\item{x}{An instance of a \linkS4class{Panel} class.}
}
\value{
For \code{.getEncodedName}, a string containing the encoded panel name of \code{x}.

For \code{.fullName}, a string containing the full (plain-English) name of the class.

For \code{.getFullName}, a string containing the full name of \code{x}.
}
\description{
Get panel names
}
\details{
The encoded name is used internally as the name of various fields in \code{input}, \code{output} and reactive lists.

The full name is what should be shown in the interface and visible to the end-user.
}
\author{
Aaron Lun
}
