% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_HeatMapPlot.R
\docType{methods}
\name{HeatMapPlot}
\alias{HeatMapPlot}
\alias{HeatMapPlot-class}
\alias{.defineParamInterface,HeatMapPlot-method}
\alias{.createParamObservers,HeatMapPlot-method}
\title{Heat map plot panel}
\description{
Create a panel with a heatmap where features are rows and samples are columns.
}
\details{
Plots reduced dimensions. What more do I have to say?
}
\section{Constructor}{

\code{HeatMapPlot()} creates an instance of a HeatMapPlot class.
}

\section{Panel parameters}{

\code{\link{.defineParamInterface}} will create parameter elements for choosing the elements to show on the heatmap.
More details to be added.
}

\examples{
#################
# For end-users #
#################

x <- HeatMapPlot()
x[["ColorScale"]]
x[["Lower"]] <- -5

##################
# For developers #
##################

library(scater)
sce <- mockSCE()
sce <- logNormCounts(sce)

# Returns NULL if there is not enough information.
sce0 <- .cacheCommonInfo(x, sce[,0])
.refineParameters(x, sce0)

# Replaces the default with something sensible.
sce0 <- .cacheCommonInfo(x, sce)
.refineParameters(x, sce0)

}
\author{
Aaron Lun
}
