% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{heatMapPlotDefaults}
\alias{heatMapPlotDefaults}
\title{Heatmap defaults}
\usage{
heatMapPlotDefaults(se, number)
}
\arguments{
\item{se}{A SummarizedExperiment object.}

\item{number}{An integer scalar, specifying the maximum number of heatmaps that can be added to the interface.}
}
\value{
A DataFrame containing default settings for parameters of each of \code{number} heatmap panels.
}
\description{
Create default settings for heatmap panels in the iSEE interface.
}
\details{
The features/rows to be used in the construction of the heatmap are specified with:
\describe{
\item{\code{FeatName}:}{List of length equal to the number of panels.
Each list entry corresponds to a panel and should be an integer vector with the indices of the feature(s) for which to show the expression in the heatmap.
Defaults to \code{1L} for each panel, i.e., the first feature in \code{se}.
Alternatively, a character vector can be supplied containing the names of the features.}
\item{\code{Assay}:}{Integer, which assay should be used to supply the expression values shown on the y-axis?
Defaults to 1, i.e., the first assay in \code{se}.
Alternatively, a string can also be supplied containing the name of the assay, if \code{assays(se)} has names.}
\item{\code{FeatNameSource}:}{Character, which other panel should be used to choose the features to show in the heatmap?
 Defaults to \code{"---"}, which means that no panel is used for feature selection.}
\item{\code{FeatNameBoxOpen}:}{Logical, should the feature selection box be open upon initialization?
Defaults to \code{FALSE}.}
}

The column metadata variables control the ordering of the samples in the heatmap.
They can be controlled with:
\describe{
\item{\code{ColData}:}{List of length equal to the number of panels.
Each list entry corresponds to a panel and should contain a character vector specifying the field(s) of \code{colData(se)} that should be used to order the samples in the heatmap.
Note that these fields will also appear as annotation bars.
Each character vector defaults to the first entry of \code{colData(se)}.}
\item{\code{ColDataBoxOpen}:}{Logical, should the column data selection box be open upon initialization?
Defaults to \code{FALSE}.}
}

A variety of parameters are available to control the color scale of the heatmap.
They can be specified with:
\describe{
\item{\code{ColorBoxOpen}:}{Logical, should the color selection panel for the heatmap be open upon initialization?
Defaults to \code{FALSE}.}
\item{\code{CenterScale}:}{List of length equal to the number of panels.
Each list entry corresponds to a panel and contains a character vector specifying whether each row of expression values should be mean-centered and/or scaled to unit variance.
Defaults to \code{"Centered"}.
Users can set it to \code{c("Centered", "Scaled")} to obtain mean-centered and unit-scaled rows.}
\item{\code{Lower}:}{Numeric, what should be the lower bound of the color scale for the values in the heatmap? All values below this threshold will be shown in the same color.
Defaults to -Inf, meaning that the lowest value in the data matrix will be used.}
\item{\code{Upper}:}{Numeric, what should be the upper bound of the color scale for the values in the heatmap? All values above this threshold will be shown in the same color.
Defaults to Inf, meaning that the highest value in the data matrix will be used.}
\item{\code{ColorScale}:}{Character, what color scale (in the form low-mid-high) should be used to color the heatmap when values are centered?
Defaults to \code{"purple-black-yellow"}.}
}

The \code{ZoomData} field for heatmaps should contain an integer vector of consecutive indices to zoom into from the full heatmap.
This vector will subset the entries in \code{FeatName} for a given panel.
This defaults to \code{NULL}, i.e., all specified features in \code{FeatName} are shown.

All plot-based parameters described in \code{?"\link{iSEE selection parameters}"} are also applicable.
}
\examples{
example(SingleCellExperiment, echo=FALSE) # mock up 'sce'.
heatMapPlotDefaults(sce, n=1)
}
\seealso{
\code{?"\link{iSEE selection parameters}"}
}
\author{
Charlotte Soneson
}
