% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentColorMap.R
\name{isColorMapCompatible}
\alias{isColorMapCompatible}
\title{Check compatibility between ExperimentColorMap and SummarizedExperiment}
\usage{
isColorMapCompatible(colormap, se, error = FALSE)
}
\arguments{
\item{colormap}{A \linkS4class{ExperimentColorMap}.}

\item{se}{An object that is coercible to \linkS4class{SingleCellExperiment}.}

\item{error}{A logical value that indicates whether an informative error
should be thrown, describing why the two objects are not compatible.}
}
\value{
A logical value that indicates whether a given pair of
ExperimentColorMap and SummarizedExperiment objects are compatible.
If \code{error=TRUE}, an informative error is thrown,
rather than returing \code{FALSE}.
}
\description{
Check compatibility between ExperimentColorMap and SummarizedExperiment
}
\examples{

# Example color maps ----

count_colors <- function(n){
  c("black","brown","red","orange","yellow")
}

qc_color_fun <- function(n){
  qc_colors <- c("forestgreen", "firebrick1")
  names(qc_colors) <- c("Y", "N")
  return(qc_colors)
}

ecm <- ExperimentColorMap(
    assays = list(
        tophat_counts = count_colors
    ),
    colData = list(
        passes_qc_checks_s = qc_color_fun
    )
)

# Example SingleCellExperiment ----

library(scRNAseq)
data(allen)
library(scater)
sce <- as(allen, "SingleCellExperiment")

# Test for compatibility ----

isColorMapCompatible(ecm, sce)

}
