% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEE-pkg.R
\docType{package}
\name{iSEE-pkg}
\alias{iSEE-pkg}
\alias{iSEE-pkg-package}
\title{iSEE: interactive SingleCell/Summarized Experiment Explorer}
\description{
iSEE does this (description)
}
\author{
Aaron Lun \email{alun@wehi.edu.au}

Charlotte Soneson \email{charlotte.soneson@uzh.ch}

Federico Marini \email{marinif@uni-mainz.de}

Kevin Rue-Albrecht \email{kevin.rue-albrecht@kennedy.ox.ac.uk}

Maintainer: John or Jane Doe \email{tobeclarified@who.does.it}
}
