% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{class}
\name{iSEEconstants-class}
\alias{iSEEconstants-class}
\title{iSEE constants}
\value{
An object of class \code{iSEEconstants}.
}
\description{
\pkg{iSEE} and extension packages use constants to represent UI choices that are not data-driven.
}
\details{
This class allows each package to store all its constants in a single object that can be exported and queried for values.
In particular, the \code{$} operator provides a mechanism to emit deprecation warnings and redirect renamed constants.
}
\section{Slots}{

\describe{
\item{\code{info}}{Read-only data.frame of information about the constants.}
}}

\examples{
constants <- new("iSEEconstants")
constants$key <- "value"
}
