% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{iSEEconstants}
\alias{iSEEconstants}
\title{iSEE Constants}
\description{
Constant values used throughout iSEE panels and extensions.
}
\section{Multiple selection parameters}{

\describe{
\item{\code{.noSelection}}{Value displayed in the absence of selection.}
}
}

\author{
Kevin Rue-Albrecht
}
