% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{.buildLabs}
\alias{.buildLabs}
\title{Generate ggplot title and label instructions}
\usage{
.buildLabs(
  x = NULL,
  y = NULL,
  color = NULL,
  shape = NULL,
  size = NULL,
  fill = NULL,
  group = NULL,
  title = NULL,
  subtitle = NULL
)
}
\arguments{
\item{x}{The character label for the horizontal axis.}

\item{y}{x The character label for the vertical axis.}

\item{color}{The character title for the color scale legend.}

\item{shape}{The character title for the point shape legend.}

\item{size}{The character title for the point size legend.}

\item{fill}{The character title for the color fill legend.}

\item{group}{The character title for the group legend.}

\item{title}{The character title for the plot title.}

\item{subtitle}{The character title for the plot subtitle}
}
\value{
Title and label instructions for \code{\link{ggplot}} as a character value.
}
\description{
Generate ggplot title and label instructions
}
\details{
If any argument is \code{NULL}, the corresponding label is not set.
}
\examples{
cat(.buildLabs(y = "Title for Y axis", color = "Color label"))
}
\seealso{
\code{\link{.scatter_plot}},
\code{\link{.violin_plot}},
\code{\link{.square_plot}}
}
\author{
Kevin Rue-Albrecht
}
