% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_commands.R
\name{manage_commands}
\alias{manage_commands}
\alias{.textEval}
\title{Manage commands to be evaluated}
\usage{
.textEval(cmd, envir)
}
\arguments{
\item{cmd}{A character vector containing commands to be executed.}

\item{envir}{An environment in which to execute the commands.}
}
\value{
\code{.textEval} returns the output of \code{eval(parse(text=cmd), envir)},
unless \code{cmd} is empty in which case it returns \code{NULL}.
}
\description{
Functions to manage commands to be evaluated.
}
\examples{
myenv <- new.env()
myenv$x <- "Hello world!"
.textEval("print(x)", myenv)
}
\author{
Aaron Lun, Kevin Rue-Albrecht
}
