% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modes.R
\name{mode_gating}
\alias{mode_gating}
\title{App pre-configured to link multiple feature expression plots}
\usage{
mode_gating(se, featExprMax = 3, features = data.frame(x =
  sample(rownames(se), size = featExprMax, replace = TRUE), y =
  sample(rownames(se), size = featExprMax, replace = TRUE)), ...,
  plot_width = 4)
}
\arguments{
\item{se}{An object that coercible to
\code{\linkS4class{SingleCellExperiment}}.}

\item{featExprMax}{Maximal number of feature expression plots in the app.}

\item{features}{\code{data.frame} with columns named \code{x} and \code{y}
that define the features on the axes of the linked plots.
Plots are serially linked from the first row to the last.}

\item{...}{Additional arguments passed to \code{\link{iSEE}}.}

\item{plot_width}{The grid width of linked plots (numeric vector of
length either 1 or equal to \code{nrow(features)}}
}
\value{
A Shiny App preconfigure with multiple chain-linked feature
expression plots is launched for interactive data exploration of the
\code{\link{SingleCellExperiment}} / \code{\link{SummarizedExperiment}}
object
}
\description{
App pre-configured to link multiple feature expression plots
}
\examples{
library(scRNAseq)
data(allen)
class(allen)

# Example data ----

library(scater)
sce <- as(allen, "SingleCellExperiment")
counts(sce) <- assay(sce, "tophat_counts")
sce <- normalize(sce)
sce <- runPCA(sce)
sce <- runTSNE(sce)
sce

# launch the app itself ----

app <- mode_gating(sce, featExprMax = 6)
if (interactive()) {
  shiny::runApp(app, port = 1234)
}
}
