% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{new_panel_file}
\alias{new_panel_file}
\title{Create a new panel file}
\usage{
new_panel_file(encoded, decoded, parent = "Panel")
}
\arguments{
\item{encoded}{Name of the new panel class.}

\item{decoded}{Extended name of the new panel class (for display).}

\item{parent}{Name of the parent panel class}
}
\description{
Opens a template R script in the editor, to define a new iSEE panel class .
}
\examples{
new_panel_file("NewRedDimPlot", "New reduced dimension plot", "RedDimPlot")
}
\seealso{
\linkS4class{Panel}
}
\author{
Kevin Rue-Albrecht
}
