% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults2.R
\docType{data}
\name{opts_panel}
\alias{opts_panel}
\title{Default panel options}
\format{An object of class \code{list} of length 4.}
\usage{
opts_panel
}
\description{
Set global options using \code{opts_panel$set()}, so that all panels will use these default values during initialization.
}
\details{
See \code{str(knitr::opts_panel$get())} for a list of default panel options.

Note that \code{opts_panel$restore()} can be used to reset the global options to the package default values.
}
\examples{
opts_chunk$get('downsample'); opts_chunk$get('selected.color')
}
\keyword{datasets}
