% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plot-generics}
\alias{plot-generics}
\alias{.generateDotPlotData}
\alias{.getCommandsPlot}
\alias{.getComandsDataColor}
\alias{.getCommandsDataShape}
\alias{.getCommandsDataSize}
\alias{.getCommandsDataFacets}
\alias{.getCommandsDataSelect}
\alias{.getCommandsPlotColor}
\title{Plotting generics}
\description{
A series of generics for controlling how plotting is performed in \linkS4class{DotPlot} panels.
Subclasses can specialize one or more of them to modify their behavior in specific panels.
}
\section{Generating plotting data}{

In \code{.generateDotPlotData(x, envir)}, the following arguments are required:
\itemize{
\item \code{x}, an instance of a Table subclass.
\item \code{envir}, the evaluation environment in which the data.frame is to be constructed.
This can be assumed to have \code{se}, the \linkS4class{SummarizedExperiment} object containing the current dataset;
possibly \code{col_selected}, if a multiple column selection is being transmitted to \code{x};
and possibly \code{row_selected}, if a multiple row selection is being transmitted to \code{x}.
}

In return, the method should add a \code{plot.data} variable in \code{envir} containing a data.frame with columns named \code{"X"} and \code{"Y"}, denoting the variables to show on the x- and y-axes respectively.
The return value should be a character vector of commands that produces \code{plot.data} when evaluated in \code{envir}.

Each row of the data.frame should correspond to one row or column in the SummarizedExperiment \code{envir$se} for \linkS4class{RowDotPlot}s and \linkS4class{ColumnDotPlot}s respectively.
Note that, even if only a subset of rows/columns in the SummarizedExperiment are to be shown, there must always be one row in the data.frame per row/column of the SummarizedExperiment, and in the same order.
All other rows of the data.frame should be filled in with \code{NA}s rather than omitted entirely.
This is necessary for correct interactions with later methods that add other variables to \code{plot.data}.

Any internal variables that are generated by the commands should be prefixed with \code{.} to avoid potential clashes with reserved variable names in the rest of the application.
}

