% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot-utils}
\alias{plot-utils}
\alias{.addFacets}
\title{Process faceting choices}
\usage{
.addFacets(x)
}
\arguments{
\item{x}{A single-row DataFrame that contains all the
input settings for the current panel.}
}
\value{
A string containing a command to define the row and column faceting
covariates.
}
\description{
Generate ggplot instructions to facet a plot by row and/or column
}
\examples{
x <- ReducedDimensionPlot(FacetByRow = "Covariate_1", FacetByColumn = "Covariate_2")
.addFacets(x)
}
\author{
Kevin Rue-Albrecht.
}
