% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{iSEE point parameters}
\alias{iSEE point parameters}
\title{Point plot aesthetic parameters}
\description{
Parameters related to aesthetics for point-based plotting panels.
}
\section{Coloring parameters}{

\describe{
\item{\code{ColorBy}:}{Character, what type of data should be used for coloring?
Defaults to \code{"None"}, but can also be \code{"Feature name"}, \code{"Sample name"}, or \code{"Column data"} (for column-based plots) or \code{"Row data"} for (row-based plots).}
\item{\code{ColorByDefaultColor}:}{String specifying the default point colour when \code{ColorBy="None"}.
Defaults to \code{"black"}.}
\item{\code{ColorByFeatName}:}{Integer, the index of the feature to use if \code{ColorBy="Feature name"}.
Defaults to 1, i.e., the first feature in \code{se}.
Alternatively, a string can be supplied containing the name of the feature.}
\item{\code{ColorBySampName}:}{Integer, the index of the sample to use if \code{ColorBy="Sample name"}.
Defaults to 1, i.e., the first sample in \code{se}.
Alternatively, a string can be supplied containing the name of the sample.}
\item{\code{ColorByRowTable}:}{Character, which row statistics table should be used to choose a feature to color by, if \code{ColorBy="Feature name"}?
Any setting will override \code{ColorByFeatName} with the selected row in the chosen table upon initialization of the app.
Defaults to \code{"---"}, which means that no table will be used.}
\item{\code{ColorByColTable}:}{Character, which column statistics table should be used to choose a sample to color by, if \code{ColorBy="Sample name"}?
Any setting will override \code{ColorBySampName} with the selected row in the chosen table upon initialization of the app.
Defaults to \code{"---"}, which means that no table will be used.}
}

For the plots where each point represents a sample (i.e., all plots except for heatmaps and row data plots), the following additional options apply:
\describe{
\item{\code{ColorByColData}:}{Character, which column of \code{colData(se)} should be used for colouring if \code{ColorBy="Column data"}?
Defaults to the first entry of \code{colData(se)}.}
\item{\code{ColorByFeatNameAssay}:}{Integer, which assay should be used to supply the expression values for colouring if \code{ColorBy="Feature name"}?
Defaults to 1, i.e., the first assay in \code{se}.
Alternatively, a string can also be supplied containing the name of the assay, if \code{assays(se)} has names.}
\item{\code{ColorBySampNameColor}:}{String specifying the colour to be used to highlight the selected sample if \code{ColorBy="Sample name"}.
Defaults to \code{"red"}.}
}

For plots where each point represents a feature (i.e., row data plots), the following additional options apply:
\describe{
\item{\code{ColorByRowData}:}{Character, which column of \code{rowData(se)} should be used for colouring if \code{ColorBy="Row data"}?
Defaults to the first entry of \code{rowData(se)}.}
\item{\code{ColorByFeatNameColor}:}{String specifying the colour to be used to highlight the selected feature if \code{ColorBy="Feature name"}.
Defaults to \code{"red"}.}
\item{\code{ColorBySampNameAssay}:}{Integer, which assay should be used to supply the expression values for colouring if \code{ColorBy="Sample name"}?
Defaults to 1, i.e., the first assay in \code{se}.
Alternatively, a string can also be supplied containing the name of the assay, if \code{assays(se)} has names.}
}
}

\section{Shape parameters}{

\describe{
\item{\code{ShapeBy}:}{Character, what type of data should be used for controlling shape?
Defaults to \code{"None"}, but can also be \code{"Column data"}.}
}

For the plots where each point represents a sample (i.e., all plots except for heatmaps and row data plots), the following additional options apply:
\describe{
\item{\code{ShapeByColData}:}{Character, which column of \code{colData(se)} should be used for shaping if \code{ShapeBy="Column data"}?
This should refer to a categorical variable, and will default to the first such entry of \code{colData(se)}.}
}

For plots where each point represents a feature (i.e., row data plots), the following additional options apply:
\describe{
\item{\code{ShapeByRowData}:}{Character, which column of \code{rowData(se)} should be used for shaping if \code{ShapeBy="Row data"}?
This should refer to a categorical variable, and will default to the first such entry of \code{rowData(se)}.}
}
}

\section{Size parameters}{

\describe{
\item{\code{SizeBy}:}{Character, what type of data should be used for controlling size?
Defaults to \code{"None"}, but can also be \code{"Column data"}.}
}

For the plots where each point represents a sample (i.e., all plots except for heatmaps and row data plots), the following additional options apply:
\describe{
\item{\code{SizeByColData}:}{Character, which column of \code{colData(se)} should be used for sizing if \code{SizeBy="Column data"}?
This should refer to a continuous variable, and will default to the first such entry of \code{colData(se)}.}
}

For plots where each point represents a feature (i.e., row data plots), the following additional options apply:
\describe{
\item{\code{SizeByRowData}:}{Character, which column of \code{rowData(se)} should be used for sizing if \code{SizeBy="Row data"}?
This should refer to a continuous variable, and will default to the first such entry of \code{rowData(se)}.}
}
}

\section{Contour line parameters}{

\describe{
\item{\code{ContourAdd}:}{Logical, should contour lines be added to the plot?}
\item{\code{ContourColor}:}{String specifying the default colour of contour lines.}
}
}

\section{Faceting parameters}{

\describe{
\item{\code{FacetByRow}:}{Logical indicating whether the plot should be faceted by row.}
\item{\code{FacetByColumn}:}{Logical indicating whether the plot should be faceted by row.}
}

For the plots where each point represents a sample (i.e., all plots except for heatmaps and row data plots), the following additional options apply:
\describe{
\item{\code{RowFacetByColData}:}{Character, which column of \code{colData(se)} should be used to facet by row?
This should refer to a categorical variable, and will default to the first such entry of \code{colData(se)}.}
\item{\code{ColumnFacetByColData}:}{Character, which column of \code{colData(se)} should be used to facet by column?
This should refer to a categorical variable, and will default to the first such entry of \code{colData(se)}.}
}

For the plots where each point represents a feature (i.e., row data plots), the following additional options apply:
\describe{
\item{\code{RowFacetByRowData}:}{Character, which column of \code{colData(se)} should be used to facet by row?
This should refer to a categorical variable, and will default to the first such entry of \code{colData(se)}.}
\item{\code{ColumnFacetByRowData}:}{Character, which column of \code{colData(se)} should be used to facet by column?
This should refer to a categorical variable, and will default to the first such entry of \code{colData(se)}.}
}
}

\section{Other plot parameters}{

Parameter visibility options are:
\describe{
\item{\code{DataBoxOpen}:}{Logical, should the data parameter box be open upon initialization?
Defaults to \code{FALSE}.}
\item{\code{VisualBoxOpen}:}{Logical, should the visual parameter box be open upon initialization?
Defaults to \code{FALSE}.}
\item{\code{VisualChoices}:}{A list containing one character vector, specifying the visual box parameters to be shown upon initialization.
This defaults to \code{list("Color")} to show only the coloring parameters, but the internal vector can also contain \code{"Points"}, \code{"Facets"} and \code{"Other"}.}
}

Options related to the appearance of points are:
\describe{
\item{\code{PointSize}:}{Numeric, the size of the points.
Defaults to 1.}
\item{\code{PointAlpha}:}{Numeric, what level of transparency should be used for the points?
Ignored when \code{SelectEffect="Transparent"} and the transmitting plot has a non-\code{NULL} selection of points.
Defaults to 1.}
\item{\code{Downsample}:}{Logical, indicating whether downsampling of overlapping points should be performed.
Defaults to \code{FALSE}.}
\item{\code{SampleRes}:}{Numeric, specifying the downsampling resolution, i.e., the granularity at which points are considered to overlap.
Higher values result in a more stringent definition of overlaps, and thus less downsampling.
Defaults to 200.}
}

Text-related parameters are:
\describe{
\item{\code{FontSize}:}{Numeric, size of the font.
Defaults to 1.}
\item{\code{LegendPosition}:}{String specifying the legend position.
Defaults to \code{"Bottom"} but can also be \code{"Right"}.}
\item{\code{ZoomData}:}{A list containing numeric vectors of length 4, containing values with names \code{"xmin"}, \code{"xmax"}, \code{"ymin"} and \code{"ymax"}.
These define the zoom window on the x- and y-axes.
Each element of the list defaults to \code{NULL}, i.e., no zooming is performed.}
}
}

\examples{
example(SingleCellExperiment, echo=FALSE) # mock up 'sce'.
redDimPlotDefaults(sce, n=1)
rowDataPlotDefaults(sce, n=1)
}
\seealso{
\code{\link{redDimPlotDefaults}},
\code{\link{featAssayPlotDefaults}},
\code{\link{colDataPlotDefaults}},
\code{\link{rowDataPlotDefaults}},
\code{\link{sampAssayPlotDefaults}}
}
\author{
Aaron Lun, Kevin Rue-Albrecht, Charlotte Soneson
}
