% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voice.R
\name{prepareSpeechRecognition}
\alias{prepareSpeechRecognition}
\title{Prepare speech recognition}
\usage{
prepareSpeechRecognition(use = FALSE)
}
\arguments{
\item{use}{Whether speech recognition should be enabled.}
}
\value{
A list of HTML content to include in the user interface.
}
\description{
Prepare speech recognition
}
\author{
Kevin Rue-Albrecht
}
