% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_RedDimPlot.R
\docType{methods}
\name{RedDimPlot}
\alias{RedDimPlot}
\alias{RedDimPlot-class}
\alias{.defineParamInterface,RedDimPlot-method}
\alias{.createParamObservers,RedDimPlot-method}
\title{The reduced dimension plot panel}
\description{
Plots reduced dimensions. What more do I have to say?
}
\section{Constructor}{

\code{RedDimPlot()} creates an instance of a RedDimPlot class.
}

\section{Panel parameters}{

\code{\link{.defineParamInterface}} will create parameter elements for choosing the reduced dimensions to plot.
More details to be added.
}

\author{
Aaron Lun
}
