% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{redDimPlotDefaults}
\alias{redDimPlotDefaults}
\title{Reduced dimension plot defaults}
\usage{
redDimPlotDefaults(se, number)
}
\arguments{
\item{se}{A SummarizedExperiment object.}

\item{number}{An integer scalar, specifying the maximum number of reduced dimension plots that can be added to the interface.}
}
\value{
A DataFrame containing default settings for parameters of each of \code{number} reduced dimension panels.
}
\description{
Create default settings for reduced dimension plot panels in the iSEE interface.
}
\details{
Parameters available to reduced dimension plots are:
\describe{
\item{\code{Type}:}{Integer, which entry of \code{reducedDims(se)} should be shown?
Defaults to 1, i.e., the first entry.
Alternatively, a string can be supplied containing the name of the reduced dimension field, if \code{reducedDims(se)} has names.}
\item{\code{XAxis}:}{Integer, which component should be shown on the x-axis?
Defaults to 1.}
\item{\code{YAxis}:}{Integer, which component should be shown on the y-axis?
Defaults to 2.}
}

All column-based parameters described in \code{?"\link{iSEE point parameters}"} are applicable.
All plot-based parameters described in \code{?"\link{iSEE selection parameters}"} are applicable.
}
\examples{
example(SingleCellExperiment, echo=FALSE) # mock up 'sce'.
redDimPlotDefaults(sce, n=1)
}
\seealso{
\code{?"\link{iSEE point parameters}"},
\code{?"\link{iSEE selection parameters}"}
}
\author{
Aaron Lun
}
