% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_class.R
\name{.removeInvalidChoices}
\alias{.removeInvalidChoices}
\title{Remove invalid values in multiple choices}
\usage{
.removeInvalidChoices(x, field, choices)
}
\arguments{
\item{x}{An instance of a \linkS4class{Panel} class.}

\item{field}{String containing the name of the relevant slot.}

\item{choices}{Character vector of permissible values for this slot.}
}
\value{
\code{x} where the slot named \code{field} is replaced only with the values that exist in \code{choices}.
}
\description{
Removes invalid values in a slot of a \linkS4class{Panel} object.
This is usually called in \code{\link{.refineParameters}}.
}
\author{
Kevin Rue-Albrecht
}
