% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reactive.R
\name{.retrieveOutput}
\alias{.retrieveOutput}
\alias{.requestUpdate}
\alias{.requestCleanUpdate}
\title{Reactive manipulations for Panel output}
\usage{
.retrieveOutput(panel_name, se, pObjects, rObjects)

.requestUpdate(panel_name, rObjects)

.requestCleanUpdate(panel_name, pObjects, rObjects)
}
\arguments{
\item{panel_name}{String containing the panel name.}

\item{se}{A \linkS4class{SummarizedExperiment} object containing the current dataset.}

\item{pObjects}{An environment containing global parameters generated in the \code{\link{iSEE}} app.}

\item{rObjects}{A reactive list of values generated in the \code{\link{iSEE}} app.}
}
\value{
\code{.retrieveOutput} will return the output of running \code{\link{.generateOutput}} for the current panel.

\code{.requestUpdate} will modify \code{rObjects} to request a re-rendering of the specified panel.
\code{.requestCleanUpdate} will also remove all active/saved selections in the chosen panel.

Both functions will invisibly return \code{NULL}.
}
\description{
Respond to or request a re-rendering of the \linkS4class{Panel} output via reactive variables.
}
\details{
\code{.retrieveOutput} should be used in the expression for rendering output, e.g., in \code{\link{.renderOutput}}.
This takes care of a number of house-keeping tasks required to satisfy \code{\link{.renderOutput}}'s requirements
with respect to updating various fields in \code{pObjects}.
It also improves efficiency by retrieving cached outputs that were used elsewhere in the app.

\code{.requestUpdate} should be used in various observers to request a re-rendering of the panel,
usually in response to user-driven parameter changes in \code{\link{.createObservers}}.

\code{.requestCleanUpdate} is usually desirable for parameter changes that invalidate existing multiple selections,
e.g., if the coordinates change in a \linkS4class{DotPlot}, existing brushes and lassos are usually not applicable.
}
\author{
Aaron Lun
}
