% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_RowDataPlot.R
\docType{methods}
\name{RowDataPlot}
\alias{RowDataPlot}
\alias{RowDataPlot-class}
\alias{.defineParamInterface,RowDataPlot-method}
\alias{.createParamObservers,RowDataPlot-method}
\title{The column data plot panel}
\description{
Plots column data values. What more do I have to say?
}
\section{Constructor}{

\code{RowDataPlot()} creates an instance of a RowDataPlot class.
}

\section{Panel parameters}{

\code{\link{.defineParamInterface}} will create parameter elements for choosing the reduced dimensions to plot.
More details to be added.
}

\author{
Aaron Lun
}
