% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{rowDataPlotDefaults}
\alias{rowDataPlotDefaults}
\title{Row data plot defaults}
\usage{
rowDataPlotDefaults(se, number)
}
\arguments{
\item{se}{A SummarizedExperiment object.}

\item{number}{An integer scalar, specifying the maximum number of row data plots that can be added to the interface.}
}
\value{
A DataFrame containing default settings for parameters of each of \code{number} row data panels.
}
\description{
Create default settings for row data plot panels in the iSEE interface.
}
\details{
Parameters available to row data plots are:
\describe{
\item{\code{YAxis}:}{Character, which column of \code{rowData(se)} should be shown on the y-axis?
Defaults to the first entry of \code{colData(se)}.}
\item{\code{XAxis}:}{Character, what variable should be shown on the x-axis?
Defaults to \code{"None"}, but can also be \code{"Row data"} or \code{"Feature name"}.}
\item{\code{XAxisRowData}:}{Character, which column of \code{rowData(se)} should be shown on the x-axis if \code{XAxis="Row data"}?
Defaults to the first entry of \code{rowData(se)}.}
}

All row-based parameters described in \code{?"\link{iSEE point parameters}"} are applicable.
All plot-based parameters described in \code{?"\link{iSEE selection parameters}"} are applicable.
}
\examples{
example(SingleCellExperiment, echo=FALSE) # mock up 'sce'.
rowDataPlotDefaults(sce, n=1)
}
\seealso{
\code{?"\link{iSEE point parameters}"},
\code{?"\link{iSEE selection parameters}"}
}
\author{
Aaron Lun
}
