% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_RowStatTable.R
\name{RowStatTable}
\alias{RowStatTable}
\alias{RowStatTable-class}
\alias{.createRenderedOutput,RowStatTable-method}
\alias{.getEncodedName,RowStatTable-method}
\title{Row statistics table panel}
\description{
Defines a \code{\link{datatable}} panel containing statistics 
from the \code{\link{rowData}} of a \linkS4class{SummarizedExperiment}.
}
\section{Constructor}{

\code{RowStatTable()} creates an instance of a RowStatTable class.
}

\section{Panel parameters}{

\code{\link{.defineParamInterface}} will create parameter elements for choosing the reduced dimensions to plot.
More details to be added.
}

\examples{
#################
# For end-users #
#################

x <- RowStatTable()
x[["Selected"]] 
x[["Selected"]] <- 2L

##################
# For developers #
##################

library(scater)
sce <- mockSCE()

# Sets the search columns appropriately. 
sce <- .cacheCommonInfo(x, sce)
.refineParameters(x, sce)

}
\author{
Aaron Lun
}
