% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{rowStatTableDefaults}
\alias{rowStatTableDefaults}
\title{Row statistics table defaults}
\usage{
rowStatTableDefaults(se, number)
}
\arguments{
\item{se}{A SummarizedExperiment object.}

\item{number}{An integer scalar, specifying the maximum number of row statistics tables that can be added to the interface.}
}
\value{
A DataFrame containing default settings for parameters of each of \code{number} row statistics table panels.
}
\description{
Create default settings for row statistics table panels in the iSEE interface.
}
\details{
Parameters available to row statistics tables are:
\describe{
\item{\code{Selected}:}{Integer, containing the index of the row to be initially selected.
Defaults to the first row, i.e., 1.
Alternatively, a string can be supplied containing the row name.}
\item{\code{Search}:}{Character, containing the initial value of the search field.
Defaults to an empty string.}
\item{\code{SearchColumns}:}{A list containing character vectors of length equal to the number of columns in \code{rowData(se)},
specifying the initial value of the search field for each column.
All entries default to an empty string.}
}

All table-based parameters described in \code{?"\link{iSEE selection parameters}"} are applicable.
}
\examples{
example(SingleCellExperiment, echo=FALSE) # mock up 'sce'.
rowStatTableDefaults(sce, n=1)
}
\seealso{
\code{?"\link{iSEE selection parameters}"}
}
\author{
Aaron Lun
}
