% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_SampAssayPlot.R
\docType{methods}
\name{SampAssayPlot}
\alias{SampAssayPlot}
\alias{SampAssayPlot-class}
\alias{.defineParamInterface,SampAssayPlot-method}
\alias{.createParamObservers,SampAssayPlot-method}
\title{The feature assay plot panel}
\description{
Plots feature assay values. What more do I have to say?
}
\section{Constructor}{

\code{SampAssayPlot()} creates an instance of a SampAssayPlot class.
}

\section{Panel parameters}{

\code{\link{.defineParamInterface}} will create parameter elements for choosing the reduced dimensions to plot.
More details to be added.
}

\author{
Aaron Lun
}
