% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{sampAssayPlotDefaults}
\alias{sampAssayPlotDefaults}
\title{Sample assay plot defaults}
\usage{
sampAssayPlotDefaults(se, number)
}
\arguments{
\item{se}{A SummarizedExperiment object.}

\item{number}{An integer scalar, specifying the maximum number of sample assay plots that can be added to the interface.}
}
\value{
A DataFrame containing default settings for parameters of each of \code{number} sample assay panels.
}
\description{
Create default settings for sample assay plot panels in the iSEE interface.
}
\details{
Parameters available to sample assay plots are:
\describe{
\item{\code{YAxis}:}{Integer, which column of \code{se} should be shown on the y-axis?
Defaults to 1, i.e., the first column.
Alternatively, a character field can be supplied containing the name of the column.}
\item{\code{Assay}:}{Integer, which assay should be used to supply the expression values shown on the y-axis?
Defaults to 1, i.e., the first assay in \code{se}.
Alternatively, a string can also be supplied containing the name of the assay, if \code{assays(se)} has names.}
\item{\code{XAxis}:}{Character, what variable should be shown on the x-axis?
Defaults to \code{"None"}, but can also be \code{"Row data"} or \code{"Sample"}.}
\item{\code{XAxisRowData}:}{Character, which column of \code{rowData(se)} should be shown on the x-axis if \code{XAxis="Row data"}?
Defaults to the first entry of \code{rowData(se)}.}
\item{\code{XAxisSample}:}{Integer, which column of \code{se} should be shown on the x-axis?
Defaults to 2 if \code{se} contains multiple columns, otherwise it is set to 1.
Alternatively, a character field can be supplied containing the name of the column.}
}

All row-based parameters described in \code{?"\link{iSEE point parameters}"} are applicable.
All plot-based parameters described in \code{?"\link{iSEE selection parameters}"} are applicable.
}
\examples{
example(SingleCellExperiment, echo=FALSE) # mock up 'sce'.
sampAssayPlotDefaults(sce, n=1)
}
\seealso{
\code{?"\link{iSEE point parameters}"},
\code{?"\link{iSEE selection parameters}"}
}
\author{
Charlotte Soneson
}
