% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{iSEE selection parameters}
\alias{iSEE selection parameters}
\title{Selection parameters}
\description{
Parameters that control the selection of transmitter plots and the effect of selection in a variety of panels.
}
\section{Selection parameters for plots}{

\describe{
\item{\code{SelectBoxOpen}:}{Logical, should the selection parameter box be open upon initialization?
Defaults to \code{FALSE}.}
\item{\code{SelectByPlot}:}{Character, which other plot should be used for point selection in the current plot?
Defaults to \code{"---"}, which means that no plot is used for point selection.}
\item{\code{SelectEffect}:}{Character, what is the effect of receiving point selection information?
Can be \code{"Restrict"}, where only the selected points are shown; \code{"Color"}, where the selected points have a different color;
or \code{"Transparent"}, where all points other than those selected are made transparent.
Defaults to \code{"Transparent"}.}
\item{\code{SelectColor}:}{Character, what color should be used for the selected points when \code{SelectEffect="Color"}?
Defaults to \code{"red"}.}
\item{\code{SelectAlpha}:}{Numeric, what level of transparency should be used for the unselected points when \code{SelectEffect="Transparent"}?
This should lie in [0, 1], where 0 is fully transparent and 1 is fully opaque.
Defaults to 0.1.}
}
For row-based plots, each point represents a feature, while for column-based plots and heatmaps, each point represents a sample.
}

\section{Inter-plot transmission rules}{

Point selection cannot occur between row-based and column-based plots.
This is because each point in a row-based plot is a feature, while each point represents a sample in the other plots.
Thus, point selection can only occur between plots of the same point type.
The only exception to this rule is the custom panels, which can receive transmissions from both row- and column-based plots -- see \code{\link{customDataPlotDefaults}}.
This is because they can generate arbitrary plots and are not truly row- or column-based.
}

\section{Selection parameters for tables}{

For the row statistics tables, the following options apply:
\describe{
\item{\code{SelectBoxOpen}:}{Logical, should the point selection parameter box be open upon initialization?
Defaults to \code{FALSE}.}
\item{\code{SelectByPlot}:}{Character, which other plot should be used to select features in the current table?
Defaults to \code{"---"}, which means that no plot is used for point selection.}
}
Only row-based plots (i.e., row data and sample assay plots) can be used for selecting points to supply to tables, for the same reasons described above.
}

\examples{
example(SingleCellExperiment, echo=FALSE) # mock up 'sce'.
redDimPlotDefaults(sce, n=1)
rowStatTableDefaults(sce, n=1)
}
\seealso{
\code{\link{redDimPlotDefaults}},
\code{\link{featAssayPlotDefaults}},
\code{\link{colDataPlotDefaults}},
\code{\link{rowDataPlotDefaults}},
\code{\link{sampAssayPlotDefaults}},
\code{\link{heatMapPlotDefaults}},
\code{\link{rowStatTableDefaults}}
}
\author{
Aaron Lun, Kevin Rue-Albrecht
}
