% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector.R
\name{selector}
\alias{selector}
\title{Interactively select points}
\usage{
selector(x, y = NULL, multiSelect = c("first", "last", "all"),
  width = 5, height = 500, ...)
}
\arguments{
\item{x}{A numeric vector of x-coordinates.}

\item{y}{A numeric vector of y-coordinates; alternatively, only \code{x} can be provided, see \code{\link{xy.coords}}.}

\item{multiSelect}{A string indicating how multiple selections should be handled.}

\item{width}{Integer scalar between 1 and 6 specifying the width of the plot in grid units.}

\item{height}{Integer scalar specifying the height of the plot in pixels.}

\item{...}{Further arguments to be passed to \code{\link{plot}}.}
}
\value{
A shiny app object.
}
\description{
Provides an interface to interactively define multiple groups of interest, possibly for use as column/row metadata in the main \code{\link{iSEE}} app.
}
\details{
This function will create an interface that is specialized for point selection.
Multiple groups can be easily selected in a single sitting, and the identity of the selected points returned upon stopping the app.
This is intended for users to create arbitrary metadata fields to use elsewhere, e.g., in the main \code{\link{iSEE}} app.

The nature of the stop-return value depends on the choice of \code{multiSelect}.
If \code{multiSelect="all"}, the return value is a list of integer vectors where each entry corresponds to one group and contains indices for all elements in that group.
Otherwise, the return value is a factor of length equal to \code{x} containing the group assignment for each element.
If \code{multiSelect="first"}, an element is assigned to the first group it is selected in, while if \code{multiSelect="last"}, it is assigned to the last group (i.e., overwrite).
}
\examples{
x <- runif(1000)
y <- rnorm(1000)
app <- selector(x, y)

if (interactive()) {
   out <- shiny::runApp(app)
}
}
\author{
Aaron Lun, based on old code in \pkg{scran}.
}
