% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_class.R
\name{.setNestedList}
\alias{.setNestedList}
\title{Set a nested list}
\usage{
.setNestedList(x, i, value)
}
\arguments{
\item{x}{A list or \code{NULL}.}

\item{i}{A character vector with at least one value, specifying the list index at each level.}

\item{value}{The value to set inside the nested list.}
}
\value{
A list is returned containing the contents of \code{x}, except that \code{x[[i]]} is set to \code{value}.
}
\description{
Set a value inside a nested list, creating intermediate lists if necessary.
}
\examples{
.setNestedList(NULL, c("A", "B", "C"), list(yay=20))
.setNestedList(list(A=list(B=3)), c("A", "B", "C"), list(yay=20))

}
\seealso{
\code{\link{.setCachedCommonInfo}} and \code{\link{setGlobalAppOptions}}, where this function is used.
}
\author{
Aaron Lun
}
