% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{single-select-generics}
\alias{single-select-generics}
\alias{.singleSelectionDimension}
\alias{.singleSelectionValue}
\title{Generics for controlling single selections}
\description{
A panel can create a single selection on either the rows or columns
and transmit this selection to another panel for use as an aesthetic parameter.
For example, users can click on a \linkS4class{RowTable} to select a gene of interest,
and then the panel can transmit the identities of that row to another panel for coloring by that selected gene's expression.
This suite of generics controls the behavior of these single selections.
}
\section{Specifying the nature of the selection}{

Given an instance of the \linkS4class{Panel} class \code{x}, \code{.singleSelectionDimension(x)} should return a string 
specifying whether the selection contains a row (\code{"row"}), 
column (\code{"column"}) or if the Panel in \code{x} does not perform single selections at all (\code{"none"}).
The output should be constant for all instances of \code{x}.
}

\section{Obtaining the selected element}{

\code{.singleSelectionValue}(x, contents) should return a string specifying the selected row or column.
If no row or column is selected, it should return \code{NULL}.

\code{contents} is any arbitrary structure set by the rendering expression in \code{\link{.renderOutput}} for \code{x}.
This should contain all of the information necessary to determine the name of the selected row/column.
For example, a data.frame of coordinates is stored by \linkS4class{DotPlot}s to identify the point selected by a brush/lasso.
}

\author{
Aaron Lun
}
