% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_Table.R
\name{Table-class}
\alias{Table-class}
\alias{initialize,Table-method}
\alias{.createObservers,Table-method}
\alias{.renderOutput,Table-method}
\alias{.defineOutput,Table-method}
\alias{.hideInterface,Table-method}
\alias{.multiSelectionCommands,Table-method}
\alias{.multiSelectionHasActive,Table-method}
\title{The Table class}
\description{
The Table is a virtual class for all panels containing a \code{\link{datatable}} widget from the \pkg{DT} package, where each row \emph{usually} corresponds to a row or column of the \linkS4class{SummarizedExperiment} object.
It provides observers for monitoring table selection, global search and column-specific search.
}
\section{Slot overview}{

The following slots control aspects of the \code{DT::datatable} interface:
\itemize{
\item \code{Selected}, a string containing the row name of the currently selected row.
Defaults to \code{NA}, in which case the value should be chosen by the subclass' \code{\link{.refineParameters}} method.
\item \code{Search}, a string containing the regular expression for the global search.
Defaults to \code{""}, i.e., no search.
\item \code{SearchColumns}, a character vector where each entry contains the search string for each column.
Defaults to an empty character vector, i.e., no search.
}

In addition, this class inherits all slots from its parent \linkS4class{Panel} class.
}

\section{Contract description}{

The Table will provide interface elements to create the \code{DT::datatble} widget.
It will also provide observers to update slots based on user interactions with the widget. 
Interface elements and observers are also provided for slots in its parent class \linkS4class{Panel}.

Subclasses are expected to implement methods for (at least) \code{\link{.getTableFunction}}.
}

\section{Supported methods}{

In the following code snippets, \code{x} is an instance of a \linkS4class{Table} class.
Refer to the documentation for each method for more details on the remaining arguments.

For defining the interface:
\itemize{
\item \code{\link{.defineOutput}(x, id)} returns a UI element for a \code{\link[DT]{dataTableOutput}} widget.
}

For defining reactive expressions:
\itemize{
\item \code{\link{.createObservers}(x, se, input, session, pObjects, rObjects)} sets up observers for all of the slots. 
This will also call the equivalent \linkS4class{Panel} method.
\item \code{\link{.renderOutput}(x, se, colormap, output, pObjects, rObjects)} will add a rendered \code{\link{datatable}} object to \code{output}.
It will also create a rendered UI element for selection information.
}

For controlling selections: 
\itemize{
\item \code{\link{.multiSelectionRestricted}(x)} returns \code{TRUE}.
Transmission of a selection to a Table will manifest as a subsetting of the rows.
\item \code{\link{.multiSelectionHasActive}(x)} returns a logical scalar indicating whether \code{x} has any active search fields.
\item \code{\link{.multiSelectionCommands}(x, index)} returns a character vector of R expressions that - when evaluated - return a character vector of the row names of the table after applying all search filters.
The value of \code{index} is ignored.
\item \code{\link{.singleSelectionValue}(x, pObjects)} returns the name of the row that was last selected in the \code{\link{datatable}} widget.
}

Unless explicitly specialized above, all methods from the parent class \linkS4class{Panel} are also available.
}

\seealso{
\linkS4class{Panel}, for the immediate parent class.
}
\author{
Aaron Lun
}
