% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tours.R
\name{tourTCGA_preprocess}
\alias{tourTCGA_preprocess}
\alias{tourTCGA}
\title{Tour of The Cancer Genome Atlas RNA-sequencing data set}
\usage{
tourTCGA_preprocess(irlba.maxit = 1000)

tourTCGA(sce)
}
\arguments{
\item{sce}{\code{SingleCellExperiment} object obtained from
\code{tourTCGA_preprocess}}
}
\value{
An instance of Shiny app that runs the tour.
}
\description{
Tour of The Cancer Genome Atlas RNA-sequencing data set
}
\examples{
sce_tour <- tourTCGA_preprocess()
app <- tourTCGA(sce_tour)
if (interactive()) {
  shiny::runApp(app, port = 1234)
}
}
