% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotateEntrez}
\alias{annotateEntrez}
\title{Annotation via Entrez database}
\usage{
annotateEntrez(se, species, keytype, rowdata_col = NULL)
}
\arguments{
\item{se}{An object that is coercible to \code{\linkS4class{SingleCellExperiment}}.}

\item{species}{A character string, indicating which species the samples are stemming from. See 
details for the format expected, and the possible values}

\item{keytype}{The keytype that matches the IDs used in the \code{se} object. Can be one of 
the values of \code{keytypes(org.XX.eg.db)}, typically being "SYMBOL", "ENSEMBL", or "ENTREZID"}

\item{rowdata_col}{A character string, corresponding to one of the columns (if present) in
\code{rowData(se)}. Defaults to NULL, which corresponds to having the ids of the features
as row names of the \code{se} object itself.}
}
\value{
A function to be used as the value of the \code{annotFun} parameter of \code{iSEE}. 
This function itself returns a \code{HTML} tag object with the content extracted from the call,
with parameteres the \code{se} object, and the \code{row_index} corresponding to the feature
of interest.
}
\description{
Annotation facility for displaying additional information on selected genes, based on
the data retrieved from the ENTREZ database
}
\details{
The \code{species} parameter has to be one of the following: Anopheles, Arabidopsis,
Bovine, Canine, Chicken, Chimp, E coli strain K12, E coli strain Sakai, Fly, Human, Malaria,                
Mouse, Pig, Rat, Rhesus, Streptomyces coelicolor, Toxoplasma gondii, Worm, Xenopus, Yeast, Zebrafish.
             
Since the functionality depends on the corresponding org.XX.eg.db packages, these
are expected to be installed, and get loaded when running the function
}
\examples{
library(scRNAseq)
data(allen)
sce <- as(allen, "SingleCellExperiment")
annotateEntrez(sce,"Mouse","SYMBOL")

# to be used when launching the app itself ----

app <- iSEE(sce)
if (interactive()) {
  shiny::runApp(app, port = 1234, annotFun = annotateEntrez(sce,"Mouse","SYMBOL"))
}
}
