% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotateEntrez}
\alias{annotateEntrez}
\title{Annotation via Entrez database}
\usage{
annotateEntrez(se, orgdb, keytype, rowdata_col = NULL)
}
\arguments{
\item{se}{An object that is coercible to \linkS4class{SingleCellExperiment}.}

\item{orgdb}{An OrgDb object, as a basis for the annotation.
Typical values are \code{org.Hs.eg.db} for human, \code{org.Mm.eg.db} for mouse, and so on.
The corresponding package has to be available and loaded before calling this function.}

\item{keytype}{The keytype that matches the IDs used in the \code{se} object.
Can be one of \code{keytypes(org.XX.eg.db)}, typically \code{"SYMBOL"}, \code{"ENSEMBL"}, or \code{"ENTREZID"}.}

\item{rowdata_col}{A character string specifying which column of \code{rowData(se)} contains the keys.
Defaults to NULL, which corresponds to having the ids of the features as row names of the \code{se} object itself.}
}
\value{
A function to be used as the value of the \code{annotFun} parameter of \code{iSEE}.
This function itself returns a \code{HTML} tag object with the content extracted from the call,
accepting as parameters the \code{se} object and the \code{row_index} corresponding to the feature of interest.
}
\description{
Annotation facility for displaying additional information on selected genes,
based on the data retrieved from the ENTREZ database
}
\examples{
library(scRNAseq)
sce <- ReprocessedAllenData(assays="tophat_counts")

library(org.Mm.eg.db)
myfun <- annotateEntrez(sce, org.Mm.eg.db, keytype="SYMBOL")
\dontrun{
# Requires a working internet connection
myfun(sce, 4242)
}

# to be used when launching the app itself ----

app <- iSEE(sce, annotFun = myfun)
if (interactive()) {
  shiny::runApp(app, port = 1234)
}
}
