% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_ColDataPlot.R
\docType{methods}
\name{ColDataPlot}
\alias{ColDataPlot}
\alias{ColDataPlot-class}
\alias{.defineParamInterface,ColDataPlot-method}
\alias{.createParamObservers,ColDataPlot-method}
\title{The column data plot panel}
\description{
Plots column data values. What more do I have to say?
}
\section{Constructor}{

\code{ColDataPlot()} creates an instance of a ColDataPlot class.
}

\section{Panel parameters}{

\code{\link{.defineParamInterface}} will create parameter elements for choosing the reduced dimensions to plot.
More details to be added.
}

\examples{
#################
# For end-users #
#################

x <- ColDataPlot()
x[["XAxis"]]
x[["XAxis"]] <- "Column data"

##################
# For developers #
##################

library(scater)
sce <- mockSCE()
sce <- logNormCounts(sce)

old_cd <- colData(sce)
colData(sce) <- NULL

# Spits out a NULL and a warning if there is nothing to plot.
sce0 <- .cacheCommonInfo(x, sce)
.refineParameters(x, sce0)

# Replaces the default with something sensible.
colData(sce) <- old_cd
sce0 <- .cacheCommonInfo(x, sce)
.refineParameters(x, sce0)

}
\author{
Aaron Lun
}
