% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_ColStatTable.R
\name{ColStatTable}
\alias{ColStatTable}
\alias{ColStatTable-class}
\alias{.createRenderedOutput,ColStatTable-method}
\alias{.getEncodedName,ColStatTable-method}
\title{Column statistics table panel}
\description{
Defines a \code{\link{datatable}} panel containing statistics 
from the \code{\link{colData}} of a \linkS4class{SummarizedExperiment}.
}
\section{Constructor}{

\code{ColStatTable()} creates an instance of a ColStatTable class.
}

\section{Panel parameters}{

\code{\link{.defineParamInterface}} will create parameter elements for choosing the reduced dimensions to plot.
More details to be added.
}

\examples{
#################
# For end-users #
#################

x <- ColStatTable()
x[["Selected"]] 
x[["Selected"]] <- 2L

##################
# For developers #
##################

library(scater)
sce <- mockSCE()

# Search column refinement works as expected.
sce0 <- .cacheCommonInfo(x, sce)
.refineParameters(x, sce0)

}
\author{
Aaron Lun
}
