% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{constants}
\alias{constants}
\alias{.dataParamBoxOpen}
\alias{.multiSelectHistory}
\alias{.noSelection}
\alias{.organizationHeight}
\alias{.organizationWidth}
\title{Constants}
\description{
Constant values used throughout iSEE panels and extensions.
The value of each constant can be accessed using the \code{$} operator and the corresponding key (listed below).
}
\section{Point colouring parameters}{

\describe{
\item{\code{colorByNothingTitle}}{Radio button choice for coloring data points by a fixed color.}
\item{\code{colorByColDataTitle}}{Radio button choice for coloring data points by column zwdata.}
\item{\code{colorByRowDataTitle}}{Radio button choice for coloring data points by row data.}
\item{\code{colorByFeatNameTitle}}{Radio button choice for highlighting a selected feature or coloring data points by assay data for that feature.}
\item{\code{colorBySampNameTitle}}{Radio button choice for highlighting a selected sample or coloring data points by assay data for that sample.}
\item{\code{colorByColSelectionsTitle}}{Radio button choice for coloring data points by an incoming column selection.}
\item{\code{colorByRowSelectionsTitle}}{Radio button choice for coloring data points by an incoming row selection.}
}
}

\section{Point shaping parameters}{
 
\describe{
\item{\code{shapeByNothingTitle}}{Radio button choice for shaping all data points as points.}
\item{\code{shapeByColDataTitle}}{Radio button choice for shaping data points by column data.}
\item{\code{shapeByRowDataTitle}}{Radio button choice for shaping data points by row data.}
}
}

\section{Point sizing parameters}{
 
\describe{
\item{\code{sizeByNothingTitle}}{Radio button choice for sizing all data points as points.}
\item{\code{sizeByColDataTitle}}{Radio button choice for sizing data points by column data.}
\item{\code{sizeByRowDataTitle}}{Radio button choice for sizing data points by row data.}
}
}

\section{Faceting parameters}{

\describe{
\item{\code{facetByNothingTitle}}{Radio button choice for disabling faceting.}
\item{\code{facetByRowDataTitle}}{Radio button choice for faceting by row data.}
\item{\code{facetByColDataTitle}}{Radio button choice for faceting by column data.}
\item{\code{facetByRowSelectionsTitle}}{Radio button choice for faceting by an incoming row selection.}
\item{\code{facetByColSelectionsTitle}}{Radio button choice for faceting by an incoming column selection.}
}
}

\section{Multiple selection parameters}{
 
\describe{
\item{\code{multiSelectSave}}{Character suffix appended to the panel identifier for the event triggered by the \code{\link{actionButton}} for adding a saved selection to memory.}
\item{\code{multiSelectDelete}}{Character suffix appended to the panel identifier for the event triggered by the \code{\link{actionButton}} for removing a saved selection from memory.}
\item{\code{noSelection}}{Dropdown choice for disabling an incoming selection.}
\item{\code{customSelection}}{Unused?}
}
}

\section{Zooming parameters}{
 
\describe{
\item{\code{zoomClick}}{Character suffix appended to the panel identifier for the event triggered by the double-click event in a \pkg{Shiny} brush for zooming.}
}
}

\section{Lasso parameters}{
 
\describe{
\item{\code{lassoClick}}{Character suffix appended to the panel identifier for the event triggered by the click event for adding a waypoint to a lasso selection.}
}
}

\section{Brush parameters}{
 
\describe{
\item{\code{brushField}}{Character suffix appended to the panel identifier for the input representing the current \pkg{Shiny} brush.}
}
}

\section{Button parameters}{
 
\describe{
\item{\code{buttonUpToDateLabel}}{Unused?}
\item{\code{buttonEmptyHistoryLabel}}{Button label in the absence of any saved selection.}
\item{\code{buttonDeleteLabel}}{Label of the button to delete the latest saved selection.}
\item{\code{buttonNoSelectionLabel}}{Label of the button for adding a saved selection to memory in the absence of an active selection.}
\item{\code{buttonSaveLabel}}{Label of the button for adding a saved selection to memory in the presence of an active selection.}
}
}

\section{Panel slot names (Deprecated)}{

TODO: Move this section to another man page when another class is implemented to store slot names and types (and default value?).
\describe{
\item{\code{iSEE::.dataParamBoxOpen}}{Name of slot that indicates whether the 'Data parameter' box is open.}
\item{\code{iSEE::.multiSelectHistory}}{Name of slot that stores the list of saved selections.}
\item{\code{iSEE::.organizationHeight}}{Name of slot that stores the panel height.}
\item{\code{iSEE::.organizationWidth}}{Name of slot that stores the panel width.}
}
}

\section{Multiple selection parameters (Deprecated)}{

TODO: Move this section to another page that documents the deprecation of "dot constants", and redirects to this page instead.
\describe{
\item{\code{iSEE::.noSelection}}{Value displayed in the absence of selection.}
}
}

\section{TODO}{
 
Move the following constants to the appropriate section

\describe{
\item{}{}
}
}

\examples{
iSEE::constants
iSEE::constants$colorByColDataTitle
}
\author{
Kevin Rue-Albrecht
}
\keyword{data}
