% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{constants}
\alias{constants}
\alias{.dataParamBoxOpen}
\alias{.multiSelectHistory}
\alias{.noSelection}
\alias{.organizationHeight}
\alias{.organizationWidth}
\title{Constants}
\description{
Constant values used throughout iSEE panels and extensions.
}
\section{Panel slot names (Deprecated)}{

\describe{
\item{\code{.dataParamBoxOpen}}{Name of slot that indicates whether the 'Data parameter' box is open.}
\item{\code{.multiSelectHistory}}{Name of slot that stores the list of saved selections.}
\item{\code{.organizationHeight}}{Name of slot that stores the panel height.}
\item{\code{.organizationWidth}}{Name of slot that stores the panel width.}
}
}

\section{Multiple selection parameters (Deprecated)}{

\describe{
\item{\code{.noSelection}}{Value displayed in the absence of selection.}
}
}

\author{
Kevin Rue-Albrecht
}
\keyword{data}
