% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{constants}
\alias{constants}
\title{Constants}
\description{
Constant values used throughout iSEE panels and extensions.
}
\section{Class slot names}{

\describe{
\item{\code{.organizationWidth}}{Name of panel class slot that stores panel width. See \linkS4class{Panel}.}
\item{\code{.organizationHeight}}{Name of panel class slot that stores panel height See \linkS4class{Panel}.}
}
}

\section{Multiple selection parameters}{

\describe{
\item{\code{.noSelection}}{Value displayed in the absence of selection.}
}
}

\author{
Kevin Rue-Albrecht
}
