% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_plots.R
\name{Custom iSEE plots}
\alias{Custom iSEE plots}
\title{Creating custom plots}
\description{
Instructions for creating custom plots for use in \code{\link{iSEE}}.
}
\section{Functions for custom column plots}{

Any function should accept the following arguments:
\itemize{
\item \code{se}, a SingleCellExperiment object.
\item \code{columns}, a character vector of column names, corresponding to a transmitted selection of points.
This will be different from \code{colnames(se)} if the current custom column plot is set to \code{"Restrict"}.
}
In return, it should produce a list containing the following elements:
\describe{
\item{\code{coordinates}:}{A data.frame with one row per entry of \code{columns}.
It should contain the fields \code{X} and \code{Y}, for the x- and y-axis values per row.
Row names must be set to \code{columns}.}
\item{\code{xlab}:}{A string specifying the x-axis label.}
\item{\code{ylab}:}{A string specifying the y-axis label.}
\item{\code{title}:}{A string specifying the plot title.}
}
}

\examples{
CUSTOM <- function(se, columns) {
    kept <- logcounts(se)[,columns]
    out <- prcomp(t(kept), rank.=2)$x
    coords <- data.frame(X=out[,1], Y=out[,2], row.names=columns)
    return(list(coordinates=coords, xlab="PC1", ylab="PC2", 
           title="PCA on selected points"))           
}

library(scRNAseq)
data(allen)
class(allen)

library(scater)
sce <- as(allen, "SingleCellExperiment")
counts(sce) <- assay(sce, "tophat_counts")
sce <- normalize(sce)
sce <- runPCA(sce)

rdp <- redDimPlotDefaults(sce, 1)
ccp <- customColPlotDefaults(sce, 1)
ccp$Function <- "PCA2"
ccp$SelectByPlot <- "Reduced dimension plot 1"
ccp$SelectEffect <- "Restrict"

app <- iSEE(sce, redDimArgs=rdp, customColArgs=ccp, 
   initialPanels=DataFrame(Name=c("Reduced dimension plot 1", 
       "Custom column plot 1")),
   customColFun=list(PCA2=CUSTOM))
}
\author{
Aaron Lun
}
