% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{customStatTableDefaults}
\alias{customStatTableDefaults}
\title{Custom statistics table defaults}
\usage{
customStatTableDefaults(se, number)
}
\arguments{
\item{se}{A SummarizedExperiment object.}

\item{number}{An integer scalar, specifying the maximum number of custom statistics tables that can be added to the interface.}
}
\value{
A DataFrame containing default settings for parameters of each of \code{number} custom statistics table panels.
}
\description{
Create default settings for custom statistics table panels in the iSEE interface.
}
\details{
Data parameters available to custom statistics tables are:
\describe{
\item{\code{DataBoxOpen}:}{Logical, should the data parameter box be open upon initialization?
Defaults to \code{FALSE}.}
\item{\code{Function}:}{String, which function should be used to generate a data.frame?
Defaults to \code{"---"}, i.e., no coordinates are generated.}
\item{\code{Arguments}:}{String with multiple lines specifying the initial arguments for the function, see \code{vignette("custom", package="iSEE")} for details.
Defaults to an empty string.}
\item{\code{VisibleArgs}:}{String with multiple lines specifying the initial arguments to be shown in the text area.
This can differ from \code{Arguments}, e.g., to list argument names without their values for users to enter.
Defaults to \code{NA}, which means that the value in \code{Arguments} will be shown.}
}

Selection parameters for custom statistics tables are:
\describe{
\item{\code{SelectBoxOpen}:}{Logical, should the selection parameter box be open upon initialization?
Defaults to \code{FALSE}.}
\item{\code{ColumnSource}:}{Character, which other plot should transmit sample selections to the current table?
Defaults to \code{"---"}, which means that no plot is used for point selection.}
\item{\code{RowSource}:}{Character, which other plot should transmit feature selections to the current table?
Defaults to \code{"---"}, which means that no plot is used for point selection.}
}

Tables also have an \code{Search} field indicating what string should be put into the search box.
This defaults to an empty string.
}
\examples{
example(SingleCellExperiment, echo=FALSE) # mock up 'sce'.
customStatTableDefaults(sce, n=1)
}
\author{
Aaron Lun, Kevin Rue-Albrecht
}
