% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_class.R
\name{dataframe-utils}
\alias{dataframe-utils}
\alias{.findAtomicFields}
\alias{.whichNumeric}
\alias{.whichGroupable}
\title{Class utilities}
\usage{
.findAtomicFields(x)

.whichGroupable(x)

.whichNumeric(x)
}
\arguments{
\item{x}{A data.frame or \linkS4class{DataFrame}.}
}
\value{
\code{.findAtomicFields}: A character vector of names of atomic fields in \code{x}.

\code{.whichGroupable}: An integer vector containing the indices of the categorical columns.

\code{.whichNumeric}: An integer vector containing the indices of the numeric columns.
}
\description{
\code{.findAtomicFields}: A utility function to find columns in a data.frame or \linkS4class{DataFrame}
that are atomic R types, as most of the app does not know how to handle  more complex types being stored as columns.
An obvious example is in data.frames expected by \code{\link{ggplot}} or \code{\link{datatable}}.
}
\details{
\code{.whichNumeric}: Identify continuous columns that can be used as options in various interface elements, e.g., for sizing.
This is typically called in \code{\link{.cacheCommonInfo}} for later use by methods of \code{\link{.defineInterface}}.

\code{.whichGroupable}: Identify categorical columns that can be used as options in various interface elements, e.g., for faceting or shaping.
This is typically called in \code{\link{.cacheCommonInfo}} for later use by methods of \code{\link{.defineInterface}}.
}
\examples{
x <- DataFrame(
    A = rnorm(10),
    B = sample(letters, 10),
    DataFrame = I(DataFrame(
        C = rnorm(10),
        D = sample(letters, 10)
    ))
)

.findAtomicFields(x)
.whichGroupable(x)
.whichNumeric(x)
}
\author{
Aaron Lun, Kevin Rue-Albrecht, Charlotte Soneson
}
