% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{defaults}
\alias{defaults}
\alias{redDimPlotDefaults}
\alias{featAssayPlotDefaults}
\alias{colDataPlotDefaults}
\alias{rowStatTableDefaults}
\alias{colStatTableDefaults}
\alias{rowDataPlotDefaults}
\alias{sampAssayPlotDefaults}
\alias{heatMapPlotDefaults}
\title{Deprecated default functions}
\usage{
redDimPlotDefaults(se, number)

featAssayPlotDefaults(se, number)

colDataPlotDefaults(se, number)

rowStatTableDefaults(se, number)

colStatTableDefaults(se, number)

rowDataPlotDefaults(se, number)

sampAssayPlotDefaults(se, number)

heatMapPlotDefaults(se, number)
}
\arguments{
\item{se}{A SummarizedExperiment object.}

\item{number}{Integer scalar specfiying the number of panels to use.}
}
\value{
A DataFrame containing default settings for parameters of each of \code{number} feature assay panels.
}
\description{
These default-creating functions are deprecated and should be replaced by constructor calls.
}
\details{
The \code{\link{iSEE}} function will attempt to translate the default parameters here into a 
the new S4 framework based on the \linkS4class{Panel} and its subclasses.
}
\examples{
example(SingleCellExperiment, echo=FALSE) # mock up 'sce'.
redDimPlotDefaults(sce, n=1)
featAssayPlotDefaults(sce, n=1)
colDataPlotDefaults(sce, n=1)

rowStatTableDefaults(sce, n=1)
colStatTableDefaults(sce, n=1)

rowDataPlotDefaults(sce, n=1)
sampAssayPlotDefaults(sce, n=1)
heatMapPlotDefaults(sce, n=1)
}
\author{
Aaron Lun
}
