% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{defaults}
\alias{defaults}
\alias{redDimPlotDefaults}
\alias{geneExprPlotDefaults}
\title{Plot parameter defaults}
\usage{
redDimPlotDefaults(se, max.plots)

geneExprPlotDefaults(se, max.plots)
}
\arguments{
\item{se}{A SingleCellExperiment object.}

\item{max.plots}{An integer scalar, specifying the maximum number of 
plots of the corresponding type that can be added to the interface.}
}
\value{
A DataFrame containing default settings for various 
parameters of reduced dimension or gene expression plots.
}
\description{
Create default settings for various plots in the iSEE interface.
}
\section{Reduced dimension plot parameters}{

\describe{
\item{\code{Active}:}{Logical, should this plot be shown upon initialization?
By default, this is only \code{TRUE} for the first plot.}
\item{\code{Type}:}{Characater, what entry of \code{reducedDims(se)} should be shown?
By default, the first entry is shown.}
\item{\code{XAxis}:}{Integer, which component should be shown on the x-axis?
Defaults to 1.}
\item{\code{YAxis}:}{Integer, which component should be shown on the y-axis?
Defaults to 2.}
\item{\code{OpenPlotPanel}:}{Logical, should the plot parameter panel be open upon initialization?
Defaults to \code{TRUE}.}
\item{\code{ColorBy}:}{Character, what type of data should be used for coloring?
Defaults to \code{"none"}.}
\item{\code{ColorByColData}:}{Character, what field of the column data should be used for colouring if \code{ColorBy="Column data"}? 
Defaults to the first entry of \code{colData(se)}.}
\item{\code{ColorByGeneExprs}:}{Character, which gene's expression should be used for colouring if \code{ColorBy="Gene expression"}? 
Defaults to the name of the first row in \code{se}.}
\item{\code{ColorByGeneExprsAssay}:}{Character, what expression values should be used for colouring if \code{ColorBy="Gene expression"}? 
Defaults to the name of the first assay in \code{se}.}
}
}

\section{Gene expression plot parameters}{

\describe{
\item{\code{Active}:}{Logical, should this plot be shown upon initialization?
By default, this is only \code{TRUE} for the first plot.}
\item{\code{Gene}:}{Character, which gene's expression values should be shown on the y-axis?
Defaults to the name of the first row in \code{se}.}
\item{\code{Assay}:}{Character, what expression values should be shown on the y-axis?
Defaults to the name of the first assay in \code{se}.}
\item{\code{XAxis}:}{Character, what variable should be shown on the x-axis?
Defaults to \code{"none"}.}
\item{\code{XAxisColData}:}{Character, what field of the column data should be shown on the x-axis if \code{XAxis="Column data"}?
Defaults to the first entry of \code{colData(se)}.}
\item{\code{XAxisGeneExprs}:}{Character, which gene's expression should be shown on the x-axis if \code{XAxis="Gene expression"}? 
Defaults to the name of the first row in \code{se}, using expression values specified in \code{Assay}.}
\item{\code{OpenPlotPanel}:}{Logical, should the plot parameter panel be open upon initialization?
Defaults to \code{TRUE}.}
\item{\code{ColorBy}:}{Character, what type of data should be used for coloring?
Defaults to \code{"none"}.}
\item{\code{ColorByColData}:}{Character, what field of the column data should be used for colouring if \code{ColorBy="Column data"}? 
Defaults to the first entry of \code{colData(se)}.}
\item{\code{ColorByGeneExprs}:}{Character, which gene's expression should be used for colouring if \code{ColorBy="Gene expression"}? 
Defaults to the name of the first row in \code{se}, using expression values specified in \code{Assay}.}
}
}

\examples{
library(scRNAseq)
data(allen)
class(allen)

library(scater)
sce <- as(allen, "SingleCellExperiment")
counts(sce) <- assay(sce, "tophat_counts")
sce <- normalize(sce)
sce <- runPCA(sce)
sce

redDimPlotDefaults(sce, max.plots=5)
geneExprPlotDefaults(sce, max.plots=5)
}
