% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{defaults}
\alias{defaults}
\alias{redDimPlotDefaults}
\alias{featAssayPlotDefaults}
\alias{colDataPlotDefaults}
\alias{customColPlotDefaults}
\alias{rowStatTableDefaults}
\alias{rowDataPlotDefaults}
\alias{sampAssayPlotDefaults}
\alias{heatMapPlotDefaults}
\title{Parameter defaults}
\usage{
redDimPlotDefaults(se, number)

featAssayPlotDefaults(se, number)

colDataPlotDefaults(se, number)

customColPlotDefaults(se, number)

rowStatTableDefaults(se, number)

rowDataPlotDefaults(se, number)

sampAssayPlotDefaults(se, number)

heatMapPlotDefaults(se, number)
}
\arguments{
\item{se}{A SummarizedExperiment object.}

\item{number}{An integer scalar, specifying the maximum number of 
panels of the corresponding type that can be added to the interface.}
}
\value{
A DataFrame containing default settings for various parameters of each panel.
}
\description{
Create default settings for various panels in the iSEE interface.
}
\section{Reduced dimension plot parameters}{

\describe{
\item{\code{Type}:}{Integer, which entry of \code{reducedDims(se)} should be shown?
Defaults to 1, i.e., the first entry.
Alternatively, a string can be supplied containing the name of the reduced dimension field, if \code{reducedDims(se)} has names.}
\item{\code{XAxis}:}{Integer, which component should be shown on the x-axis?
Defaults to 1.}
\item{\code{YAxis}:}{Integer, which component should be shown on the y-axis?
Defaults to 2.}
}
}

\section{Feature assay plot parameters}{

\describe{
\item{\code{YAxisFeatName}:}{Integer, the index of the feature for which to show the expression on the y-axis if \code{YAxis="Feature name"}. 
Defaults to 1, i.e., the first feature in \code{se}.
Alternatively, a string can be supplied containing the name of the feature, i.e., the row name.}
\item{\code{YAxisRowTable}:}{Character, what row statistic table should be used to choose a feature to display on the y-axis?
Any setting will override \code{YAxisFeatName} upon initialization of the app.
Defaults to \code{"---"}, which means that no table will be used.}
\item{\code{Assay}:}{Integer, which assay should be used to supply the expression values shown on the y-axis?
Defaults to 1, i.e., the first assay in \code{se}.
Alternatively, a string can also be supplied containing the name of the assay, if \code{assays(se)} has names.}
\item{\code{XAxis}:}{Character, what type of variable should be shown on the x-axis?
Defaults to \code{"None"}, but can also be \code{"Row table"}, \code{"Column data"} or \code{"Feature name"}.}
\item{\code{XAxisColData}:}{Character, what column of \code{colData(se)} should be shown on the x-axis if \code{XAxis="Column data"}?
Defaults to the first entry of \code{colData(se)}.}
\item{\code{XAxisFeatName}:}{Integer, the index of the feature for which to show the expression on the x-axis if \code{XAxis="Feature name"}. 
Defaults to 1, i.e., the first feature in \code{se}.
Alternatively, a string can be supplied containing the name of the feature.}
\item{\code{XAxisRowTable}:}{Character, which row statistic table should be used to choose a feature to put on the x-axis if \code{XAxis="Row table"}? 
Any setting will override \code{XAxisFeatName} upon initialization of the app.
Defaults to \code{"---"}, which means that no table will be used.}
}
}

\section{Column data plot parameters}{

\describe{
\item{\code{YAxis}:}{Character, which column of \code{colData(se)} should be shown on the y-axis?
Defaults to the first entry of \code{colData(se)}.}
\item{\code{XAxis}:}{Character, what type of variable should be shown on the x-axis?
Defaults to \code{"None"}, but can also be \code{"Column data"}.}
\item{\code{XAxisColData}:}{Character, which column of \code{colData(se)} should be shown on the x-axis if \code{XAxis="Column data"}?
Defaults to the first entry of \code{colData(se)}.}
}
}

\section{Row data plot parameters}{

\describe{
\item{\code{YAxis}:}{Character, which column of \code{rowData(se)} should be shown on the y-axis?
Defaults to the first entry of \code{colData(se)}.}
\item{\code{XAxis}:}{Character, what variable should be shown on the x-axis?
Defaults to \code{"None"}, but can also be \code{"Row data"} or \code{"Feature name"}.}
\item{\code{XAxisRowData}:}{Character, which column of \code{rowData(se)} should be shown on the x-axis if \code{XAxis="Row data"}?
Defaults to the first entry of \code{rowData(se)}.}
}
}

\section{Sample assay plot parameters}{

\describe{
\item{\code{YAxis}:}{Integer, which column of \code{se} should be shown on the y-axis?
Defaults to 1, i.e., the first column.
Alternatively, a character field can be supplied containing the name of the column.}
\item{\code{Assay}:}{Integer, which assay should be used to supply the expression values shown on the y-axis?
Defaults to 1, i.e., the first assay in \code{se}.
Alternatively, a string can also be supplied containing the name of the assay, if \code{assays(se)} has names.}
\item{\code{XAxis}:}{Character, what variable should be shown on the x-axis?
Defaults to \code{"None"}, but can also be \code{"Row data"} or \code{"Sample"}.}
\item{\code{XAxisRowData}:}{Character, which column of \code{rowData(se)} should be shown on the x-axis if \code{XAxis="Row data"}?
Defaults to the first entry of \code{rowData(se)}.}
\item{\code{XAxisSample}:}{Integer, which column of \code{se} should be shown on the x-axis?
Defaults to 2 if \code{se} contains multiple columns, otherwise it is set to 1.
Alternatively, a character field can be supplied containing the name of the column.}
}
}

\section{Custom column plot parameters}{

\describe{
\item{\code{Function}:}{String, which function should be used to generate coordinates for each sample?
Defaults to \code{"---"}, i.e., no coordinates are generated.}
}
}

\section{Coloring parameters (for point-based plots)}{

\describe{
\item{\code{ColorBy}:}{Character, what type of data should be used for coloring?
Defaults to \code{"None"}, but can also be \code{"Feature name"} or \code{"Column data"} (for column-based plots) or \code{"Row data"} for (row-based plots).}
\item{\code{ColorByDefaultColor}:}{String specifying the default point colour when \code{ColorBy="None"}.
Defaults to \code{"black"}.}
\item{\code{ColorByFeatName}:}{Integer, the index of the feature to use for colouring based on expression, if \code{ColorBy="Feature name"}? 
Defaults to 1, i.e., the first feature in \code{se}.
Alternatively, a string can be supplied containing the name of the feature.}
\item{\code{ColorByRowTable}:}{Character, which row statistic table should be used to choose a feature to color by, if \code{ColorBy="Feature name"}? 
Any setting will override \code{ColorByFeatName} upon initialization of the app.
Defaults to \code{"---"}, which means that no table will be used.}
}

For the plots where each point represents a sample (i.e., all plots except for heatmaps and row data plots), the following additional options apply:
\describe{
\item{\code{ColorByColData}:}{Character, which column of \code{colData(se)} should be used for colouring if \code{ColorBy="Column data"}? 
Defaults to the first entry of \code{colData(se)}.}
\item{\code{ColorByFeatNameAssay}:}{Integer, which assay should be used to supply the expression values for colouring if \code{ColorBy="Feature name"}? 
Defaults to 1, i.e., the first assay in \code{se}.
Alternatively, a string can also be supplied containing the name of the assay, if \code{assays(se)} has names.}
}

For plots where each point represents a feature (i.e., row data plots), the following additional options apply:
\describe{
\item{\code{ColorByRowData}:}{Character, which column of \code{rowData(se)} should be used for colouring if \code{ColorBy="Row data"}? 
Defaults to the first entry of \code{rowData(se)}.}
\item{\code{ColorByFeatNameColor}:}{String specifying the colour to be used to highlight the selected feature from the text if \code{ColorBy="Feature name"}.
Defaults to \code{"red"}.}
}
}

\section{Shape parameters (for point-based plots)}{

\describe{
\item{\code{ShapeBy}:}{Character, what type of data should be used for controlling shape?
Defaults to \code{"None"}, but can also be \code{"Column data"}.}
}

For the plots where each point represents a sample (i.e., all plots except for heatmaps and row data plots), the following additional options apply:
\describe{
\item{\code{ShapeByColData}:}{Character, which column of \code{colData(se)} should be used for shaping if \code{ShapeBy="Column data"}? 
This should refer to a categorical variable, and will default to the first such entry of \code{colData(se)}.}
}

For plots where each point represents a feature (i.e., row data plots), the following additional options apply:
\describe{
\item{\code{ShapeByRowData}:}{Character, which column of \code{rowData(se)} should be used for shaping if \code{ShapeBy="Row data"}? 
This should refer to a categorical variable, and will default to the first such entry of \code{rowData(se)}.}
}
}

\section{Contour line parameters (for scatter plots)}{

\describe{
\item{\code{ContourAdd}:}{Logical, should contour lines be added to the plot?}
\item{\code{ColorByDefaultColorContour}:}{String specifying the default colour of contour lines.}
}
}

\section{Faceting parameters (for point-based plots)}{

\describe{
\item{\code{FacetByRow}:}{Logical indicating whether the plot should be faceted by row.}
\item{\code{FacetByColumn}:}{Logical indicating whether the plot should be faceted by row.}
}

For the plots where each point represents a sample (i.e., all plots except for heatmaps and row data plots), the following additional options apply:
\describe{
\item{\code{RowFacetByColData}:}{Character, which column of \code{colData(se)} should be used to facet by row?
This should refer to a categorical variable, and will default to the first such entry of \code{colData(se)}.}
\item{\code{ColumnFacetByColData}:}{Character, which column of \code{colData(se)} should be used to facet by column?
This should refer to a categorical variable, and will default to the first such entry of \code{colData(se)}.}
}

For the plots where each point represents a feature (i.e., row data plots), the following additional options apply:
\describe{
\item{\code{RowFacetByRowData}:}{Character, which column of \code{colData(se)} should be used to facet by row?
This should refer to a categorical variable, and will default to the first such entry of \code{colData(se)}.}
\item{\code{ColumnFacetByRowData}:}{Character, which column of \code{colData(se)} should be used to facet by column?
This should refer to a categorical variable, and will default to the first such entry of \code{colData(se)}.}
}
}

\section{Other plot parameters (for point-based plots)}{

Parameter visibility options are:
\describe{
\item{\code{DataBoxOpen}:}{Logical, should the data parameter box be open upon initialization?
Defaults to \code{FALSE}.}
\item{\code{VisualBoxOpen}:}{Logical, should the visual parameter box be open upon initialization?
Defaults to \code{FALSE}.}
\item{\code{VisualChoices}:}{A list containing one character vector, specifying the visual box parameters to be shown upon initialization.
This defaults to \code{list("Color")} to show only the coloring parameters, but the internal vector can also contain \code{"Points"} and \code{"Other"}.}
}

Options related to the appearance of points are:
\describe{
\item{\code{PointSize}:}{Numeric, the size of the points.
Defaults to 1.}
\item{\code{PointAlpha}:}{Numeric, what level of transparency should be used for the points?
Ignored when \code{SelectEffect="Transparent"} and the transmitting plot has a non-\code{NULL} selection of points.
Defaults to 1.}
\item{\code{Downsample}:}{Logical, indicating whether downsampling of overlapping points should be performed.
Defaults to \code{FALSE}.}
\item{\code{SampleRes}:}{Numeric, specifying the downsampling resolution, i.e., the granularity at which points are considered to overlap.
Higher values result in a more stringent definition of overlaps, and thus less downsampling.
Defaults to 200.}
}

Text-related parameters are:
\describe{
\item{\code{FontSize}:}{Numeric, size of the font.
Defaults to 1.}
\item{\code{LegendPosition}:}{String specifying the legend position.
Defaults to \code{"Bottom"} but can also be \code{"Right"}.}
\item{\code{ZoomData}:}{A list containing numeric vectors of length 4, containing values with names \code{"xmin"}, \code{"xmax"}, \code{"ymin"} and \code{"ymax"}.
These define the zoom window on the x- and y-axes. 
Each element of the list defaults to \code{NULL}, i.e., no zooming is performed.}
}
}

\section{Row statistics table parameters}{

\describe{
\item{\code{Selected}:}{Integer, containing the index of the row to be initially selected.
Defaults to the first row, i.e., 1.
Alternatively, a string can be supplied containing the row name.}
\item{\code{Search}:}{Character, containing the initial value of the search field.
Defaults to an empty string.}
\item{\code{SearchColumns}:}{A list containing character vectors of length equal to the number of columns in \code{rowData(se)},
specifying the initial value of the search field for each column.
All entries default to an empty string.}
}
}

\section{Heatmap parameters}{

The features/rows to be used in the construction of the heatmap are specified with:
\describe{
\item{\code{FeatName}:}{List of length equal to the number of panels.
Each list entry corresponds to a panel and should be an integer vector with the indices of the feature(s) for which to show the expression in the heatmap. 
Defaults to \code{1L} for each panel, i.e., the first feature in \code{se}.
Alternatively, a character vector can be supplied containing the names of the features.}
\item{\code{Assay}:}{Integer, which assay should be used to supply the expression values shown on the y-axis?
Defaults to 1, i.e., the first assay in \code{se}.
Alternatively, a string can also be supplied containing the name of the assay, if \code{assays(se)} has names.}
\item{\code{FeatNameSource}:}{Character, which other panel should be used to choose the features to show in the heatmap?
 Defaults to \code{"---"}, which means that no panel is used for feature selection.}
\item{\code{FeatNameBoxOpen}:}{Logical, should the feature selection box be open upon initialization?
Defaults to \code{FALSE}.}
}

The column metadata variables control the ordering of the samples in the heatmap. 
They can be controlled with:
\describe{
\item{\code{ColData}:}{List of length equal to the number of panels.
Each list entry corresponds to a panel and should contain a character vector specifying the field(s) of \code{colData(se)} that should be used to order the samples in the heatmap.
Note that these fields will also appear as annotation bars.
Each character vector defaults to the first entry of \code{colData(se)}.}
\item{\code{ColDataBoxOpen}:}{Logical, should the column data selection box be open upon initialization?
Defaults to \code{FALSE}.}
}

A variety of parameters are available to control the color scale of the heatmap.
They can be specified with:
\describe{ 
\item{\code{ColorBoxOpen}:}{Logical, should the color selection panel for the heatmap be open upon initialization?
Defaults to \code{FALSE}.}
\item{\code{CenterScale}:}{List of length equal to the number of panels.
Each list entry corresponds to a panel and contains a character vector specifying whether each row of expression values should be mean-centered and/or scaled to unit variance.
Defaults to \code{"Centered"}. 
Users can set it to \code{c("Centered", "Scaled")} to obtain mean-centered and unit-scaled rows.}
\item{\code{Lower}:}{Numeric, what should be the lower bound of the color scale for the values in the heatmap? All values below this threshold will be shown in the same color.
Defaults to -Inf, meaning that the lowest value in the data matrix will be used.}
\item{\code{Upper}:}{Numeric, what should be the upper bound of the color scale for the values in the heatmap? All values above this threshold will be shown in the same color.
Defaults to Inf, meaning that the highest value in the data matrix will be used.}
\item{\code{ColorScale}:}{Character, what color scale (in the form low-mid-high) should be used to color the heatmap when values are centered?
Defaults to \code{"purple-black-yellow"}.}
}

Finally, the \code{ZoomData} field for heatmaps should contain an integer vector of consecutive indices to zoom into from the full heatmap. 
This vector will subset the entries in \code{FeatName} for a given panel. 
This defaults to \code{NULL}, i.e., all specified features in \code{FeatName} are shown.
}

\section{Point selection parameters}{

For the point-based plots, the following options apply:
\describe{
\item{\code{SelectBoxOpen}:}{Logical, should the point selection parameter box be open upon initialization?
Defaults to \code{FALSE}.}
\item{\code{SelectByPlot}:}{Character, which other plot should be used for point selection in the current plot? 
Defaults to \code{"---"}, which means that no plot is used for point selection.}
\item{\code{SelectEffect}:}{Character, what is the effect of receiving point selection information?
Can be \code{"Restrict"}, where only the selected points are shown; \code{"Color"}, where the selected points have a different color; 
or \code{"Transparent"}, where all points other than those selected are made transparent. 
Defaults to \code{"Transparent"}.}
\item{\code{SelectColor}:}{Character, what color should be used for the selected points when \code{SelectEffect="Color"}?
Defaults to \code{"red"}.}
\item{\code{SelectAlpha}:}{Numeric, what level of transparency should be used for the unselected points when \code{SelectEffect="Transparent"}?
This should lie in [0, 1], where 0 is fully transparent and 1 is fully opaque. 
Defaults to 0.1.}
}

Note that point selection cannot occur between row data plots and the other plot types.
This is because each point in a row data plot is a feature, while each point represents a sample in the other plots.
Point selection can only occur between plots of the same point type.

For the row statistics tables, the following options apply:
\describe{
\item{\code{SelectBoxOpen}:}{Logical, should the point selection parameter box be open upon initialization?
Defaults to \code{FALSE}.}
\item{\code{SelectByPlot}:}{Character, which other plot should be used to select features in the current table? 
Defaults to \code{"---"}, which means that no plot is used for point selection.}
}
Only row data plots can be used for selecting points to supply to tables.
}

\examples{
library(scRNAseq)
data(allen)
class(allen)

library(scater)
sce <- as(allen, "SingleCellExperiment")
counts(sce) <- assay(sce, "tophat_counts")
sce <- normalize(sce)
sce <- runPCA(sce)
sce

redDimPlotDefaults(sce, number=5)
featAssayPlotDefaults(sce, number=5)
colDataPlotDefaults(sce, number=5)
rowStatTableDefaults(sce, number=5)
rowDataPlotDefaults(sce, number=5)
heatMapPlotDefaults(sce, number=5)
}
