% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{defaults}
\alias{defaults}
\alias{redDimPlotDefaults}
\alias{geneExprPlotDefaults}
\alias{colDataPlotDefaults}
\alias{geneStatTableDefaults}
\title{Parameter defaults}
\usage{
redDimPlotDefaults(se, number)

geneExprPlotDefaults(se, number)

colDataPlotDefaults(se, number)

geneStatTableDefaults(se, number)
}
\arguments{
\item{se}{A SingleCellExperiment object.}

\item{number}{An integer scalar, specifying the maximum number of 
panels of the corresponding type that can be added to the interface.}
}
\value{
A DataFrame containing default settings for various parameters of each panel.
}
\description{
Create default settings for various panels in the iSEE interface.
}
\section{Reduced dimension plot parameters}{

\describe{
\item{\code{Type}:}{Character, what entry of \code{reducedDims(se)} should be shown?
By default, the first entry is shown.}
\item{\code{XAxis}:}{Integer, which component should be shown on the x-axis?
Defaults to 1.}
\item{\code{YAxis}:}{Integer, which component should be shown on the y-axis?
Defaults to 2.}
}
}

\section{Gene expression plot parameters}{

\describe{
\item{\code{GeneTable}:}{Character, what gene statistic table should be used to choose a gene to display on the y-axis?
Defaults to an empty string, which means that the first available table will be used.}
\item{\code{Assay}:}{Character, what expression values should be shown on the y-axis?
Defaults to the name of the first assay in \code{se}.}
\item{\code{XAxis}:}{Character, what variable should be shown on the x-axis?
Defaults to \code{"None"}.}
\item{\code{XAxisColData}:}{Character, what column of \code{colData(se)} should be shown on the x-axis if \code{XAxis="Column data"}?
Defaults to the first entry of \code{colData(se)}.}
\item{\code{XAxisGeneText}:}{Character, which gene's expression should be shown on the x-axis if \code{XAxis="Gene text"}? 
Defaults to the name of the first row in \code{se}, using expression values specified in \code{Assay}.}
\item{\code{XAxisGeneTable}:}{Character, which gene statistic table should be used to choose a gene to put on the x-axis if \code{XAxis="Gene table"}? 
Defaults to an empty string, which means that the first available table will be used.}
}
}

\section{Column data plot parameters}{

\describe{
\item{\code{YAxisColData}:}{Character, which column of \code{colData(se)} should be shown on the y-axis?
Defaults to the first entry of \code{colData(se)}.}
\item{\code{XAxis}:}{Character, what variable should be shown on the x-axis?
Defaults to \code{"None"}.}
\item{\code{XAxisColData}:}{Character, which column of \code{colData(se)} should be shown on the x-axis if \code{XAxis="Column data"}?
Defaults to the first entry of \code{colData(se)}.}
}
}

\section{Coloring parameters}{

\describe{
\item{\code{ColorPanelOpen}:}{Logical, should the color parameter panel be open upon initialization?
Defaults to \code{FALSE}.}
\item{\code{ColorBy}:}{Character, what type of data should be used for coloring?
Defaults to \code{"None"}.}
\item{\code{ColorByColData}:}{Character, which column of \code{colData(se)} should be used for colouring if \code{ColorBy="Column data"}? 
Defaults to the first entry of \code{colData(se)}.}
\item{\code{ColorByGeneTable}:}{Character, which gene statistic table should be used to choose a gene to color by, if \code{ColorBy="Gene table"}? 
Defaults to an empty string, which means that the first available table will be used.}
\item{\code{ColorByGeneTableAssay}:}{Character, what expression values should be used for colouring if \code{ColorBy="Gene table"}? 
Defaults to the name of the first assay in \code{se}.}
\item{\code{ColorByGeneText}:}{Character, which gene should be used to choose a gene to color by, if \code{ColorBy="Gene text"}? 
Defaults to an empty string, which means that the first available table will be used.}
\item{\code{ColorByGeneTextAssay}:}{Character, what expression values should be used for colouring if \code{ColorBy="Gene text"}? 
Defaults to the name of the first assay in \code{se}.}
}
}

\section{Brushing parameters}{

\describe{
\item{\code{BrushPanelOpen}:}{Logical, should the brushing parameter panel be open upon initialization?
Defaults to \code{FALSE}.}
\item{\code{BrushOn}:}{Logical, should the plot be transmitting its brush for use in other plots?
Defaults to \code{FALSE}.}
\item{\code{BrushByPlot}:}{Character, which other plot should be used for point selection in the current plot? 
Defaults to an empty string, which means that no plot is used for point selection.}
\item{\code{BrushEffect}:}{Character, what is the effect of receiving a brush input?
Can be \code{"Restrict"}, where only the brushed points are shown; \code{"Color"}, where the brushed points have a different color; 
or \code{"Transparent"}, where all points other than the brushed points are made transparent. Defaults to \code{"Transparent"}.}
\item{\code{BrushColor}:}{Character, what color should be used for the brushed points when \code{BrushEffect="Color"}?
Defaults to \code{"red"}.}
\item{\code{BrushAlpha}:}{Numeric, what level of transparency should be used for the unbrushed points whe \code{BrushEffect="Transparent"}?
This should lie in [0, 1], where 0 is fully transparent and 1 is fully opaque. 
Defaults to 0.1.}
}
}

\section{Other plot parameters}{

\describe{
\item{\code{PlotPanelOpen}:}{Logical, should the plot parameter panel be open upon initialization?
Defaults to \code{FALSE}.}
\item{\code{ZoomOn}:}{Logical, should zooming (via brushing and double-clicking) be turned on?
Defaults to \code{TRUE}.}
\item{\code{ZoomData}:}{A list containing numeric vectors of length 4, containing values with names \code{"xmin"}, \code{"xmax"}, \code{"ymin"} and \code{"ymax"}.
These define the zoom window on the x- and y-axes.
Each element of the list defaults to \code{NULL}, i.e., no zooming is performed.}
}
}

\section{Gene statistic table parameters}{

\describe{
\item{\code{Selected}:}{Integer, containing the index of the row to be initially selected.
Defaults to the first row, i.e., 1.}
\item{\code{Search}:}{Character, containing the initial value of the search field.
Defaults to an empty string.}
}
}

\examples{
library(scRNAseq)
data(allen)
class(allen)

library(scater)
sce <- as(allen, "SingleCellExperiment")
counts(sce) <- assay(sce, "tophat_counts")
sce <- normalize(sce)
sce <- runPCA(sce)
sce

redDimPlotDefaults(sce, number=5)
geneExprPlotDefaults(sce, number=5)
colDataPlotDefaults(sce, number=5)
geneStatTableDefaults(sce, number=5)
}
