% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_DotPlot.R
\name{DotPlot-class}
\alias{DotPlot-class}
\alias{initialize,DotPlot-method}
\alias{.defineOutput,DotPlot-method}
\alias{.generateOutput,DotPlot-method}
\alias{.generateDotPlot,DotPlot-method}
\alias{.renderOutput,DotPlot-method}
\alias{.exportOutput,DotPlot-method}
\alias{.refineParameters,DotPlot-method}
\alias{.cacheCommonInfo,DotPlot-method}
\alias{.createObservers,DotPlot-method}
\alias{.hideInterface,DotPlot-method}
\alias{.multiSelectionRestricted,DotPlot-method}
\alias{.multiSelectionActive,DotPlot-method}
\alias{.multiSelectionCommands,DotPlot-method}
\alias{.multiSelectionClear,DotPlot-method}
\alias{.multiSelectionDimension,DotPlot-method}
\alias{.singleSelectionValue,DotPlot-method}
\alias{.singleSelectionSlots,DotPlot-method}
\alias{.prioritizeDotPlotData,DotPlot-method}
\alias{.colorByNoneDotPlotField,DotPlot-method}
\alias{.colorByNoneDotPlotScale,DotPlot-method}
\alias{.defineVisualTextInterface,DotPlot-method}
\alias{.defineVisualOtherInterface,DotPlot-method}
\title{The DotPlot virtual class}
\description{
The DotPlot is a virtual class for all panels where each row or column in the \linkS4class{SummarizedExperiment} is represented by no more than one point (i.e., a \dQuote{dot}) in a brushable \link{ggplot} plot.
It provides slots and methods to control various aesthetics of the dots and to store the brush or lasso selection.
}
\section{Slot overview}{

The following slots are relevant to coloring of the points:
\itemize{
\item \code{ColorBy}, a string specifying how points should be colored.
This should be one of \code{"None"}, \code{"Feature name"}, \code{"Sample name"} and either \code{"Column data"} (for \linkS4class{ColumnDotPlot}s) or \code{"Row data"} (for \linkS4class{RowDotPlot}s).
Defaults to \code{"None"}.
\item \code{ColorByDefaultColor}, a string specifying the default color to use for all points if \code{ColorBy="None"}.
Defaults to \code{"black"}.
\item \code{ColorByFeatureName}, a string specifying the feature to be used for coloring points when \code{ColorBy="Feature name"}.
For \linkS4class{RowDotPlot}s, this is used to highlight the point corresponding to the selected feature;
for \linkS4class{ColumnDotPlot}s, this is used to color each point according to the expression of that feature.
If \code{NA}, this defaults to the name of the first row.
\item \code{ColorByFeatureSource}, a string specifying the name of the panel to use for transmitting the feature selection to \code{ColorByFeatureName}.
Defaults to \code{"---"}.
\item \code{ColorBySampleName}, a string specifying the sample to be used for coloring points when \code{ColorBy="Sample name"}.
For \linkS4class{RowDotPlot}s, this is used to color each point according to the expression of that sample;
for \linkS4class{ColumnDotPlot}s, this is used to highlight the point corresponding to the selected sample.
If \code{NA}, this defaults to the name of the first column.
\item \code{ColorBySampleSource}, a string specifying the name of the panel to use for transmitting the sample selection to \code{ColorBySampleNameColor}.
Defaults to \code{"---"}.
\item \code{ColorByFeatureDynamicSource}, a logical scalar indicating whether \code{x} should dynamically change its selection source when coloring by feature.
Defaults to \code{FALSE}.
\item \code{ColorBySampleDynamicSource}, a logical scalar indicating whether \code{x} should dynamically change its selection source when coloring by feature.
Defaults to \code{FALSE}.
}

The following slots control other metadata-related aesthetic aspects of the points:
\itemize{
\item \code{ShapeBy}, a string specifying how the point shape should be determined.
This should be one of \code{"None"} and either \code{"Column data"} (for \linkS4class{ColumnDotPlot}s) or \code{"Row data"} (for \linkS4class{RowDotPlot}s).
Defaults to \code{"None"}.
\item \code{SizeBy}, a string specifying the metadata field for controlling point size.
This should be one of \code{"None"} and either \code{"Column data"} (for \linkS4class{ColumnDotPlot}s) or \code{"Row data"} (for \linkS4class{RowDotPlot}s).
Defaults to \code{"None"}.
}

The following slots control the faceting:
\itemize{
\item \code{FacetByRow}, a string specifying the metadata field to use for creating row facets.
For \linkS4class{RowDotPlot}s, this should be a field in the \code{\link{rowData}},
while for \linkS4class{ColumnDotPlot}s, this should be a field in the \code{\link{colData}}.
Defaults to \code{"---"}, i.e., no row faceting.
\item \code{FacetByColumn}, a string specifying the metadata field to use for creating column facets.
For \linkS4class{RowDotPlot}s, this should be a field in the \code{\link{rowData}},
while for \linkS4class{ColumnDotPlot}s, this should be a field in the \code{\link{colData}}.
Defaults to \code{"---"}, i.e., no column faceting.
}

The following slots control the effect of the transmitted selection from another panel:
\itemize{
\item \code{SelectionEffect}, a string specifying the selection effect.
This should be one of \code{"Transparent"} (the default), where all non-selected points become transparent;
\code{"Color"}, where all selected points change to the specified color;
\code{"Restrict"}, where all non-selected points are not plotted.
\item \code{SelectionAlpha}, a numeric scalar in [0, 1] specifying the transparency to use for non-selected points when \code{SelectionEffect="Transparent"}.
Defaults to 0.1.
\item \code{SelectionColor}, a string specifying the color to use for selected points when \code{SelectionEffect="Color"}.
Defaults to \code{"red"}.
}

The following slots control the behavior of brushes:
\itemize{
\item \code{ZoomData}, a named numeric vector of plot coordinates with \code{"xmin"}, \code{"xmax"}, \code{"ymin"} and \code{"ymax"} elements parametrizing the zoom boundaries.
Defaults to an empty vector, i.e., no zoom.
\item \code{BrushData}, a list containing either a Shiny brush (see \code{?\link{brushedPoints}}) or an \pkg{iSEE} lasso (see \code{?\link{lassoPoints}}).
Defaults to an empty list, i.e., no brush or lasso.
}

The following slots control some aspects of the user interface:
\itemize{
\item \code{DataBoxOpen}, a logical scalar indicating whether the data parameter box should be open.
Defaults to \code{FALSE}.
\item \code{VisualBoxOpen}, a logical scalar indicating whether the visual parameter box should be open.
Defaults to \code{FALSE}.
\item \code{VisualChoices}, a character vector specifying the visible interface elements upon initialization.
This can contain zero or more of \code{"Color"}, \code{"Shape"}, \code{"Size"}, \code{"Point"} , \code{"Facet"}, \code{"Text"}, and \code{"Other"}.
Defaults to \code{"Color"}.
}

The following slots control the addition of a contour:
\itemize{
\item \code{ContourAdd}, logical scalar indicating whether a contour should be added to a (scatter) plot.
Defaults to \code{FALSE}.
\item \code{ContourColor}, string specifying the color to use for the contour lines.
Defaults to \code{"blue"}.
}

The following slots control the general appearance of the points.
\itemize{
\item \code{PointSize}, positive numeric scalar specifying the relative size of the points.
Defaults to 1.
\item \code{PointAlpha}, non-negative numeric scalar specifying the transparency of the points.
Defaults to 1, i.e., not transparent.
\item \code{Downsample}, logical scalar indicating whether to downsample points for faster plotting.
Defaults to \code{FALSE}.
\item \code{DownsampleResolution}, numeric scalar specifying the resolution of the downsampling grid (see \code{?\link{subsetPointsByGrid}}) if \code{Downsample=TRUE}.
Larger values correspond to reduced downsampling at the cost of plotting speed.
Defaults to 200.
}

The following slots refer to general plotting parameters:
\itemize{
\item \code{FontSize}, positive numeric scalar specifying the relative font size.
Defaults to 1.
\item \code{PointSize}, positive numeric scalar specifying the relative point size.
Defaults to 1.
\item \code{LegendPosition}, string specifying the position of the legend on the plot.
Defaults to \code{"Right"} but can also be \code{"Bottom"}.
}

In addition, this class inherits all slots from its parent \linkS4class{Panel} class.
}

\section{Supported methods}{

In the following code snippets, \code{x} is an instance of a \linkS4class{DotPlot} class.
Refer to the documentation for each method for more details on the remaining arguments.

For setting up the objects:
\itemize{
\item \code{\link{.cacheCommonInfo}(x)} adds a \code{"DotPlot"} entry containing \code{valid.assay.names}, a character vector of valid assay names.
Valid names are defined as those that are non-empty, i.e., not \code{""}.
This method will also call the equivalent \linkS4class{Panel} method.
\item \code{\link{.refineParameters}(x, se)} replaces \code{NA} values in \code{ColorByFeatureName} and \code{ColorBySampleNameColor} with the first row and column name, respectively, of \code{se}.
This will also call the equivalent \linkS4class{Panel} method.
}

For defining the interface:
\itemize{
\item \code{\link{.defineOutput}(x, id)} returns a UI element for a brushable plot.
}

For generating the output:
\itemize{
\item \code{\link{.generateOutput}(x, se, all_memory, all_contents)} returns a list containing \code{contents}, a data.frame with one row per point currently present in the plot;
\code{plot}, a \link{ggplot} object;
and \code{commands}, a list of character vector containing the R commands required to generate \code{contents} and \code{plot}.
\item \code{\link{.generateDotPlot}(x, labels, envir)} returns a list containing \code{plot} and \code{commands}, as described above.
This is called within \code{\link{.generateOutput}} for all \linkS4class{DotPlot} instances by default.
Methods are also guaranteed to generate a \code{dot.plot} variable in \code{envir} containing the \link{ggplot} object corresponding to \code{plot}.
\item \code{\link{.prioritizeDotPlotData}(x, envir)} returns \code{NULL}.
\item \code{\link{.colorByNoneDotPlotField}(x)} returns \code{NULL}.
\item \code{\link{.colorByNoneDotPlotScale}(x)} returns \code{NULL}.
\item \code{\link{.exportOutput}(x, se, all_memory, all_contents)} will create a PDF file containing the current plot, and return a string containing the path to that PDF.
This assumes that the \code{plot} field returned by \code{\link{.generateOutput}} is a \link{ggplot} object.
}

For defining reactive expressions:
\itemize{
\item \code{\link{.createObservers}(x, se, input, session, pObjects, rObjects)} sets up observers for some (but not all!) of the slots.
This will also call the equivalent \linkS4class{Panel} method.
\item \code{\link{.renderOutput}(x, se, output, pObjects, rObjects)} will add a rendered plot element to \code{output}.
The reactive expression will add the contents of the plot to \code{pObjects$contents} and the relevant commands to \code{pObjects$commands}.
This will also call the equivalent \linkS4class{Panel} method to render the panel information testboxes.
}

For controlling selections:
\itemize{
\item \code{\link{.multiSelectionRestricted}(x)} returns a logical scalar indicating whether \code{x} is restricting the plotted points to those that were selected in a transmitting panel, i.e., is \code{SelectionEffect="Restrict"}?
\item \code{\link{.multiSelectionCommands}(x, index)} returns a character vector of R expressions that - when evaluated - returns a character vector of the names of selected points in the active and/or saved selections of \code{x}.
The active selection is returned if \code{index=NA}, otherwise one of the saved selection is returned.
\item \code{\link{.multiSelectionActive}(x)} returns \code{x[["BrushData"]]} or \code{NULL} if there is no brush or closed lasso.
\item \code{\link{.multiSelectionClear}(x)} returns \code{x} after setting the \code{BrushData} slot to an empty list.
\item \code{\link{.singleSelectionValue}(x)} returns the name of the first selected element in the active brush.
If no brush is active, \code{NULL} is returned instead.
\item \code{\link{.singleSelectionSlots}(x)} will return a list specifying the slots that can be updated by single selections in transmitter panels, mostly related to the choice of coloring parameters.
This includes the output of \code{callNextMethod}.
}

Unless explicitly specialized above, all methods from the parent class \linkS4class{Panel} are also available.
}

\section{Subclass expectations}{

The DotPlot is a rather vaguely defined class for which the only purpose is to avoid duplicating code for \linkS4class{ColumnDotPlot}s and \linkS4class{RowDotPlot}s.
We recommend extending those subclasses instead.
}

\seealso{
\linkS4class{RowDotPlot} and \linkS4class{ColumnDotPlot}, which are more amenable to extension.
}
\author{
Aaron Lun
}
