% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{ExperimentColorMap}
\alias{ExperimentColorMap}
\alias{class:ExperimentColorMap}
\alias{ExperimentColorMap-class}
\alias{assayColorMap}
\alias{colDataColorMap}
\alias{rowDataColorMap}
\alias{assayColorMap,ExperimentColorMap,character-method}
\alias{assayColorMap,ExperimentColorMap,numeric-method}
\alias{colDataColorMap,ExperimentColorMap,character-method}
\alias{rowDataColorMap,ExperimentColorMap,character-method}
\title{\code{ExperimentColorMap} objects}
\usage{
ExperimentColorMap(assays = list(), colData = list(), rowData = list(),
  all_discrete = list(assays = NULL, colData = NULL, rowData = NULL),
  all_continuous = list(assays = NULL, colData = NULL, rowData = NULL),
  global_discrete = NULL, global_continuous = NULL, ...)
}
\arguments{
\item{assays}{List of color maps for \code{assays}.}

\item{colData}{List of color maps for \code{colData}.}

\item{rowData}{List of color maps for \code{rowData}.}

\item{all_discrete}{Discrete color maps applied to all undefined
\code{assays}, \code{colData}, and \code{rowData}, respectively.}

\item{all_continuous}{Continuous color maps applied to all undefined
\code{assays}, \code{colData}, and \code{rowData}, respectively.}

\item{global_discrete}{Discrete color maps applied to all undefined discrete
covariates.}

\item{global_continuous}{Continuous color maps applied to all undefined
discrete covariates.}

\item{...}{additional arguments passed on to the \code{ExperimentColorMap}
constructor}
}
\value{
An object of class \code{ExperimentColorMap}
}
\description{
\code{ExperimentColorMap} objects
}
\details{
Color maps must all be functions that take at least one argument: the number
of (named) colours to return as a \code{character} vector.
This argument may be ignored in the body of the color map function
to produce constant color maps.
}
\section{Accessors}{


In the following code snippets, \code{x} is an
\code{ExperimentColorMap} object. If the color map can not immediately
be found in the appropriate slot, \code{discrete} is a \code{logical(1)}
that indicates whether the default color map returned should be discrete
\code{TRUE} or continuous (\code{FALSE}, default).

\describe{
  \item{\code{assayColorMap(x, i, ..., discrete=FALSE)}:}{
  Get an \code{assays} colormap.}
  
  \item{\code{colDataColorMap(x, i, ..., discrete=FALSE)}:}{
  Get a \code{colData} colormap.}
  
  \item{\code{rowDataColorMap(x, i, ..., discrete=FALSE)}:}{
  Get a \code{rowData} colormap.}
}
}

\examples{

# Example color maps ----

count_colors <- function(n){
  c("black","brown","red","orange","yellow")
}
fpkm_colors <- viridis::inferno
tpm_colors <- viridis::plasma

qc_color_fun <- function(n){
  qc_colors <- c("forestgreen", "firebrick1")
  names(qc_colors) <- c("Y", "N")
  return(qc_colors)
}

# Constructor ----

ecm <- ExperimentColorMap(
    assays = list(
        counts = count_colors,
        tophat_counts = count_colors,
        cufflinks_fpkm = fpkm_colors,
        cufflinks_fpkm = fpkm_colors,
        rsem_tpm = tpm_colors
    ),
    colData = list(
        passes_qc_checks_s = qc_color_fun
    )
)

# Accessors ----

assayColorMap(ecm, "logcounts") # [undefined --> default]
assayColorMap(ecm, "counts")
assayColorMap(ecm, "cufflinks_fpkm")

colDataColorMap(ecm, "passes_qc_checks_s")
colDataColorMap(ecm, "undefined")

rowDataColorMap(ecm, "undefined")

}
