% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{ExperimentColorMap}
\alias{ExperimentColorMap}
\alias{class:ExperimentColorMap}
\alias{ExperimentColorMap-class}
\alias{assayColorMap}
\alias{colDataColorMap}
\alias{rowDataColorMap}
\alias{assayColorMap,ExperimentColorMap,character-method}
\alias{assayColorMap,ExperimentColorMap,integer-method}
\alias{colDataColorMap,ExperimentColorMap,character-method}
\alias{colDataColorMap,ExperimentColorMap,integer-method}
\alias{rowDataColorMap,ExperimentColorMap,character-method}
\alias{rowDataColorMap,ExperimentColorMap,integer-method}
\title{\code{ExperimentColorMap} objects}
\usage{
ExperimentColorMap(assays = list(), colData = list(), rowData = list(),
  ...)
}
\arguments{
\item{assays}{List of color maps for \code{assays}.}

\item{colData}{List of color maps for \code{colData}.}

\item{rowData}{List of color maps for \code{rowData}.}

\item{...}{additional arguments passed on to the \code{ExperimentColorMap}
constructor}
}
\description{
\code{ExperimentColorMap} objects
}
\section{Accessors}{


In the following code snippets, \code{x} is an
\code{ExperimentColorMap} object.

\code{assayColorMap(x, "counts")}: Get an assay colormap.
}

\examples{

# Constructor ----

count_colors <- viridis::viridis(10)
logcounts_colors <- viridis::magma(10)
fpkm_colors <- viridis::inferno(10)
tpm_colors <- viridis::plasma(10)

qc_colors <- c("forestgreen", "firebrick1")
names(qc_colors) <- c("Y", "N")

ecm <- new("ExperimentColorMap",
    assays = list(
        counts = count_colors,
        tophat_counts = count_colors,
        cufflinks_fpkm = logcounts_colors,
        cufflinks_fpkm = fpkm_colors,
        rsem_tpm = tpm_colors
    ),
    colData = list(
        passes_qc_checks_s = qc_colors
    )
)

# Accessor ----

assayColorMap(ecm, "logcounts") # viridis::magma(10)
assayColorMap(ecm, "undefined") # default: viridis::viridis(10)

colDataColorMap(ecm, "passes_qc_checks_s")
colDataColorMap(ecm, "undefined")

rowDataColorMap(ecm, "undefined")

}
