% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_FeatAssayPlot.R
\docType{methods}
\name{FeatAssayPlot}
\alias{FeatAssayPlot}
\alias{FeatAssayPlot-class}
\alias{.defineParamInterface,FeatAssayPlot-method}
\alias{.createParamObservers,FeatAssayPlot-method}
\title{The feature assay plot panel}
\description{
Plots feature assay values. What more do I have to say?
}
\section{Constructor}{

\code{FeatAssayPlot()} creates an instance of a FeatAssayPlot class.
}

\section{Panel parameters}{

\code{\link{.defineParamInterface}} will create parameter elements for choosing the reduced dimensions to plot.
More details to be added.
}

\examples{
#################
# For end-users #
#################

x <- FeatAssayPlot()
x[["XAxis"]]
x[["Assay"]] <- "logcounts"
x[["XAxisColData"]] <- "stuff"

##################
# For developers #
##################

library(scater)
sce <- mockSCE()
sce <- logNormCounts(sce)

old_assay_names <- assayNames(sce)
assayNames(sce) <- character(length(old_assay_names))

# Spits out a NULL and a warning if no assays are named.
sce0 <- .cacheCommonInfo(x, sce)
.refineParameters(x, sce0)

# Replaces the default with something sensible.
assayNames(sce) <- old_assay_names
sce0 <- .cacheCommonInfo(x, sce)
.refineParameters(x, sce0)

}
\author{
Aaron Lun
}
