% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEE-pkg.R
\docType{package}
\name{iSEE-pkg}
\alias{iSEE-pkg}
\alias{iSEE-pkg-package}
\title{iSEE: interactive SummarizedExperiment/SingleCellExperiment Explorer}
\description{
\code{iSEE} is a Bioconductor package that provides an interactive Shiny-based
 graphical user interface for exploring data stored in \code{SummarizedExperiment}
 objects, including row- and column-level metadata. 
 Particular attention is given to single-cell data in a \code{SingleCellExperiment}
 object with visualization of dimensionality reduction results, e.g., from principal
 components analysis (PCA) or t-distributed stochastic neighbour embedding (t-SNE)
}
\author{
Aaron Lun \email{alun@wehi.edu.au}

Charlotte Soneson \email{charlotte.soneson@uzh.ch}

Federico Marini \email{marinif@uni-mainz.de}

Kevin Rue-Albrecht \email{kevin.rue-albrecht@kennedy.ox.ac.uk}

Maintainer: John or Jane Doe \email{tobeclarified@who.does.it}
}
