% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEE-pkg.R
\docType{package}
\name{iSEE-pkg}
\alias{iSEE-package}
\alias{iSEE-pkg}
\title{iSEE: interactive SummarizedExperiment/SingleCellExperiment Explorer}
\description{
\code{iSEE} is a Bioconductor package that provides an interactive Shiny-based
 graphical user interface for exploring data stored in \code{SummarizedExperiment}
 objects, including row- and column-level metadata. 
 Particular attention is given to single-cell data in a \code{SingleCellExperiment}
 object with visualization of dimensionality reduction results, e.g., from principal
 components analysis (PCA) or t-distributed stochastic neighbour embedding (t-SNE)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/iSEE/iSEE}
  \item Report bugs at \url{https://github.com/iSEE/iSEE/issues}
}

}
\author{
Aaron Lun \email{infinite.monkeys.with.keyboards@gmail.com}

Charlotte Soneson \email{charlottesoneson@gmail.com}

Federico Marini \email{marinif@uni-mainz.de}

Kevin Rue-Albrecht \email{kevinrue67@gmail.com}
}
