% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEE.R
\name{iSEE}
\alias{iSEE}
\title{iSEE: interactive SingleCell/Summarized Experiment Explorer}
\usage{
iSEE(se, redDim.default = 5, geneExpr.default = 5)
}
\arguments{
\item{se}{A SingleCellExperiment object.}

\item{redDim.default}{An integer scalar specifying the maximum number of
reduced dimension plots in the interface. Alternatively, a DataFrame 
similar to that produced by \code{\link{redDimPlotDefault}}, specifying 
initial parameters for the plots.}

\item{geneExpr.default}{An integer scalar specifying the maximum number of
gene expression plots in the interface. Alternatively, a DataFrame 
similar to that produced by \code{\link{geneExprPlotDefault}}, specifying
initial parameters for the plots.}
}
\value{
A Shiny App is launched for interactive data exploration of the
\code{\link{SingleCellExperiment}} / \code{\link{SummarizedExperiment}}
object
}
\description{
Interactive visualization of single-cell data using a Shiny interface.
}
\details{
Users can pass default parameters via DataFrame objects in 
\code{redDim.default} and \code{geneExpr.default}. Each object can contain 
some or all of the expected fields (see \code{\link{redDimPlotDefault}}).
Any missing fields will be filled in with the defaults.

Users can specify any number of maximum plots, though increasing the 
number will increase the time required to generate any given plot.
}
\examples{
library(scRNAseq)
data(allen)
class(allen)

library(scater)
sce <- as(allen, "SingleCellExperiment")
counts(sce) <- assay(sce, "tophat_counts")
sce <- normalize(sce)
sce <- runPCA(sce)
sce

# launch the app itself
if (interactive()) { iSEE(sce) }
}
