% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEE.R
\name{iSEE}
\alias{iSEE}
\title{iSEE: interactive SingleCell/Summarized Experiment Explorer}
\usage{
iSEE(se, redDimArgs = 5, colDataArgs = 5, geneExprArgs = 5,
  initialPanels = NULL, annot.orgdb = NULL, annot.keytype = "ENTREZID",
  annot.keyfield = NULL)
}
\arguments{
\item{se}{A \linkS4class{SingleCellExperiment} object.}

\item{redDimArgs}{An integer scalar specifying the maximum number of
reduced dimension plots in the interface. Alternatively, a DataFrame
similar to that produced by \code{\link{redDimPlotDefaults}}, specifying
initial parameters for the plots.}

\item{colDataArgs}{An integer scalar specifying the maximum number of
column data plots in the interface. Alternatively, a DataFrame
similar to that produced by \code{\link{colDataPlotDefaults}}, specifying
initial parameters for the plots.}

\item{geneExprArgs}{An integer scalar specifying the maximum number of
gene expression plots in the interface. Alternatively, a DataFrame
similar to that produced by \code{\link{geneExprPlotDefaults}}, specifying
initial parameters for the plots.}

\item{initialPanels}{A DataFrame specifying which panels should be created
at initialization. This should contain a \code{Name} character field and
a \code{Width} integer field, see Details.}

\item{annot.orgdb}{An \code{org.*.db} annotation object from which
Entrez identifiers can be retrieved.}

\item{annot.keytype}{A string specifying the keytype to use to query
\code{annot.orgdb}.}

\item{annot.keyfield}{A string specifying the field of \code{rowData(se)}
containing the keys of type \code{annot.keytype}. If \code{NULL}, the
row names of \code{se} are used as the keys.}
}
\value{
A Shiny App is launched for interactive data exploration of the
\code{\link{SingleCellExperiment}} / \code{\link{SummarizedExperiment}}
object
}
\description{
Interactive visualization of single-cell data using a Shiny interface.
}
\details{
Users can pass default parameters via DataFrame objects in
\code{redDim.args} and \code{geneExpr.args}. Each object can contain
some or all of the expected fields (see \code{\link{redDimPlotDefaults}}).
Any missing fields will be filled in with the defaults.

The number of maximum plots for each type of plot is implicitly inferred
from the number of rows of the corresponding DataFrame in \code{*Args},
if an integer scalar was not supplied. Users can specify any number of
maximum plots, though increasing the number will increase the time
required to render the interface.

The \code{initialPanels} argument specifies the panels to be created
upon initializing the interface. This should be a DataFrame containing
a \code{Name} field specifying the identity of the panel, e.g.,
\code{"Reduced dimension plot 1"}, \code{"Gene statistics table 2"}.
The trailing number should not be greater than the number of
maximum plots of that type. The \code{Width} field may also be specified
describing the width of the panel from 4 to 12 (values will be coerced
inside this range).

If \code{annot.orgdb} is specified, gene information will be retrieved
upon selection of particular genes in the data table. No retrieval is
performed if \code{annot.orgdb=NULL}.
}
\examples{
library(scRNAseq)
data(allen)
class(allen)

library(scater)
sce <- as(allen, "SingleCellExperiment")
counts(sce) <- assay(sce, "tophat_counts")
sce <- normalize(sce)
sce <- runPCA(sce)
sce <- runTSNE(sce)
sce

# launch the app itself
if (interactive()) { iSEE(sce) }
}
