% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEE-main.R
\name{iSEE}
\alias{iSEE}
\title{iSEE: interactive SummarizedExperiment Explorer}
\usage{
iSEE(
  se,
  initial = NULL,
  extra = NULL,
  redDimArgs = NULL,
  colDataArgs = NULL,
  featAssayArgs = NULL,
  rowStatArgs = NULL,
  rowDataArgs = NULL,
  sampAssayArgs = NULL,
  colStatArgs = NULL,
  customDataArgs = NULL,
  customStatArgs = NULL,
  heatMapArgs = NULL,
  redDimMax = 5,
  colDataMax = 5,
  featAssayMax = 5,
  rowStatMax = 5,
  rowDataMax = 5,
  sampAssayMax = 5,
  colStatMax = 5,
  customDataMax = 5,
  customStatMax = 5,
  heatMapMax = 5,
  initialPanels = NULL,
  annotFun = NULL,
  customDataFun = NULL,
  customStatFun = NULL,
  customSendAll = FALSE,
  colormap = ExperimentColorMap(),
  landingPage = createLandingPage(),
  tour = NULL,
  appTitle = NULL,
  runLocal = TRUE,
  voice = FALSE,
  bugs = FALSE,
  ...
)
}
\arguments{
\item{se}{An object that is coercible to \linkS4class{SingleCellExperiment}.
If missing, an app is launched with a landing page generated by the \code{landingPage} argument.}

\item{initial}{A list of \linkS4class{Panel} objects specifying the initial state of the app.
The order of panels determines the sequence in which they are laid out in the interface.}

\item{extra}{A list of additional \linkS4class{Panel} objects that might be added after the app has started.
Only the first instance of each class will be used.}

\item{redDimArgs}{Deprecated, use \code{initial} instead.}

\item{colDataArgs}{Deprecated, use \code{initial} instead.}

\item{featAssayArgs}{Deprecated, use code{initial} instead.}

\item{rowStatArgs}{Deprecated, use code{initial} instead.}

\item{rowDataArgs}{Deprecated, use code{initial} instead.}

\item{sampAssayArgs}{Deprecated, use code{initial} instead.}

\item{colStatArgs}{Deprecated, use code{initial} instead.}

\item{customDataArgs}{Deprecated, use code{initial} instead.}

\item{customStatArgs}{Deprecated, use code{initial} instead.}

\item{heatMapArgs}{Deprecated, use code{initial} instead.}

\item{redDimMax}{Deprecated and ignored.}

\item{colDataMax}{Deprecated and ignored.}

\item{featAssayMax}{Deprecated and ignored.}

\item{rowStatMax}{Deprecated and ignored.}

\item{rowDataMax}{Deprecated and ignored.}

\item{sampAssayMax}{Deprecated and ignored.}

\item{colStatMax}{Deprecated and ignored.}

\item{customDataMax}{Deprecated and ignored.}

\item{customStatMax}{Deprecated and ignored.}

\item{heatMapMax}{Deprecated and ignored.}

\item{initialPanels}{Deprecated, use \code{initial} instead.}

\item{annotFun}{Deprecated and ignored.
Read the book for the new, more extensible approach for creating custom panels.}

\item{customDataFun}{Deprecated and ignored.
Read the book for the new, more extensible approach for creating custom panels.}

\item{customStatFun}{Deprecated and ignored.
Read the book for the new, more extensible approach for creating custom panels.}

\item{customSendAll}{Deprecated and ignored.}

\item{colormap}{An \linkS4class{ExperimentColorMap} object that defines custom colormaps to apply to individual \code{assays}, \code{colData} and \code{rowData} covariates.}

\item{landingPage}{A function that renders a landing page when \code{se} is started without any specified \code{se}.}

\item{tour}{A data.frame with the content of the interactive tour to be displayed after starting up the app.}

\item{appTitle}{A string indicating the title to be displayed in the app.
If not provided, the app displays the version info of \code{\link{iSEE}}.}

\item{runLocal}{A logical indicating whether the app is to be run locally or remotely on a server, which determines how documentation will be accessed.}

\item{voice}{A logical indicating whether the voice recognition should be enabled.}

\item{bugs}{Set to \code{TRUE} to enable the bugs Easter egg.
Alternatively, a named numeric vector control the respective number of each bug type (e.g., \code{c(bugs=3L, spiders=1L)}).}

\item{...}{Further arguments to pass to \code{\link{shinyApp}}.}
}
\value{
A Shiny app object is returned for interactive data exploration of \code{se},
either by simply printing the object or by explicitly running it with \code{\link{runApp}}.
}
\description{
Interactive and reproducible visualization of data contained in a
\linkS4class{SummarizedExperiment} object, using a Shiny interface.
}
\details{
Configuring the initial state of the app is as easy as passing a list of \linkS4class{Panel} objects to \code{initial}.
Each element represents one panel and is typicall constructed with a command like \code{\link{RedDimPlot}()}.
Panels are filled from left to right in a row-wise manner depending on the available width.
Each panel can be easily customized by modifying the parameters in each object.

The \code{extra} argument should specify \linkS4class{Panel} classes that are not present in \code{initial}
but can be added interactively by the user after the app has started.
The first instance of each new class in \code{extra} will be used as a template when the user adds a new panel of that class.
Note that \code{extra} will be combined with the classes already present in \code{initial},
so any redundant specification of those initial classes will be ignored.

The \code{tour} argument needs to be provided in a form compatible with the format expected by the \code{rintrojs} package.
There should be two columns, \code{element} and \code{intro}, with the former describing the element to highlight and the latter providing some descriptive text.
See \url{https://github.com/carlganz/rintrojs#usage} for more information.

By default, categorical data types such as factor and character are limited to 24 levels, beyond which they are coerced to numeric variables for faster plotting.
This limit may be set to a different value as a global option, e.g. \code{options(iSEE.maxlevels=30)}.

The default landing page allows users to upload their own RDS files to initialize the app.
By default, the maximum request size for file uploads defaults to 5MB 
(\url{https://shiny.rstudio.com/reference/shiny/0.14/shiny-options.html}).
To raise the limit (e.g., 50MB), run \code{options(shiny.maxRequestSize=50*1024^2)}.
See \code{\link{createLandingPage}} for details on creating a custom landing page.
}
\examples{
library(scRNAseq)

# Example data ----
sce <- ReprocessedAllenData(assays="tophat_counts")
class(sce)

library(scater)
sce <- logNormCounts(sce, exprs_values="tophat_counts")

sce <- runPCA(sce, ncomponents=4)
sce <- runTSNE(sce)
rowData(sce)$ave_count <- rowMeans(assay(sce, "tophat_counts"))
rowData(sce)$n_cells <- rowSums(assay(sce, "tophat_counts") > 0)
sce

# launch the app itself ----

app <- iSEE(sce)
if (interactive()) {
  shiny::runApp(app, port=1234)
}

}
\references{
Rue-Albrecht K, Marini F, Soneson C, Lun ATL.
iSEE: Interactive SummarizedExperiment Explorer
\emph{F1000Research} 7.

Javascript code for \code{bugs} was based on \url{https://github.com/Auz/Bug}.
}
