% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEE.R
\name{iSEE}
\alias{iSEE}
\title{iSEE: interactive SingleCell/Summarized Experiment Explorer}
\usage{
iSEE(se = NULL)
}
\arguments{
\item{se}{somethingaboutit}
}
\value{
A Shiny App is launched for interactive data exploration of the
\code{\link{SingleCellExperiment}} / \code{\link{SummarizedExperiment}}
object
}
\description{
iSEE does this
}
\details{
and does that in detail
}
\examples{
library(SingleCellExperiment)
library(scRNAseq)
data(allen)
class(allen)
sce <- as(allen, "SingleCellExperiment")
sce
# launch the app itself
# iSEE(sce)
}
