% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEE-main.R
\name{iSEE}
\alias{iSEE}
\title{iSEE: interactive SummarizedExperiment/SingleCellExperiment Explorer}
\usage{
iSEE(se, redDimArgs = NULL, colDataArgs = NULL, featExprArgs = NULL,
  rowStatArgs = NULL, rowDataArgs = NULL, heatMapArgs = NULL,
  redDimMax = 5, colDataMax = 5, featExprMax = 5, rowStatMax = 5,
  rowDataMax = 5, heatMapMax = 5, initialPanels = NULL, annotFun = NULL,
  colormap = ExperimentColorMap(), tour = NULL, run_local = TRUE)
}
\arguments{
\item{se}{An object that is coercible to \linkS4class{SingleCellExperiment}.}

\item{redDimArgs}{A DataFrame similar to that produced by
\code{\link{redDimPlotDefaults}}, specifying initial parameters
for the plots.}

\item{colDataArgs}{A DataFrame similar to that produced by
\code{\link{colDataPlotDefaults}}, specifying initial parameters
for the plots.}

\item{featExprArgs}{A DataFrame similar to that produced by
\code{\link{featExprPlotDefaults}}, specifying initial parameters
for the plots.}

\item{rowStatArgs}{A DataFrame similar to that produced by
\code{\link{rowStatTableDefaults}}, specifying initial parameters
for the plots.}

\item{rowDataArgs}{A DataFrame similar to that produced by
\code{\link{rowDataPlotDefaults}}, specifying initial parameters
for the plots.}

\item{heatMapArgs}{A DataFrame similar to that produced by
\code{\link{heatMapPlotDefaults}}, specifying initial parameters
for the plots.}

\item{redDimMax}{An integer scalar specifying the maximum number of
reduced dimension plots in the interface.}

\item{colDataMax}{An integer scalar specifying the maximum number of
column data plots in the interface.}

\item{featExprMax}{An integer scalar specifying the maximum number of
feature expression plots in the interface.}

\item{rowStatMax}{An integer scalar specifying the maximum number of
row statistics tables in the interface.}

\item{rowDataMax}{An integer scalar specifying the maximum number of
row data plots in the interface.}

\item{heatMapMax}{An integer scalar specifying the maximum number of
heatmaps in the interface.}

\item{initialPanels}{A DataFrame specifying which panels should be
created at initialization. 
This should contain a \code{Name} character field and a \code{Width}
integer field, see Details.}

\item{annotFun}{A function, constructed in the form of
\code{\link{annotateEntrez}} or \code{\link{annotateEnsembl}}.
This function is built in a way to generate itself 
a function supposed to accept two parameters,
\code{se} and \code{row_index},
to have a unified yet flexible interface to generate additional information
to display for the selected genes of interest.}

\item{colormap}{An \linkS4class{ExperimentColorMap} object that defines
custom color maps to apply to individual \code{assays}, \code{colData},
and \code{rowData} covariates.}

\item{tour}{A data.frame with the content of the interactive tour to be 
displayed after starting up the app. Defaults to \code{NULL}. More 
information is provided in the details.}

\item{run_local}{A logical indicating whether the app is to be run locally
or remotely on a server, which determines how documentation will be
accessed.}
}
\value{
A Shiny App is launched for interactive data exploration of the
\code{\link{SummarizedExperiment}}/\code{\link{SingleCellExperiment}} 
object
}
\description{
Interactive and reproducible visualization of data contained in a
SummarizedExperiment/SingleCellExperiment, using a Shiny interface.
}
\details{
Users can pass default parameters via DataFrame objects in
\code{redDimArgs} and \code{featExprArgs}. Each object can contain
some or all of the expected fields (see \code{\link{redDimPlotDefaults}}).
Any missing fields will be filled in with the defaults.

The number of maximum plots for each type of plot is set to the larger
of \code{*Max} and \code{nrow(*Args)}. Users can specify any number of
maximum plots, though increasing the number will increase the time
required to render the interface.

The \code{initialPanels} argument specifies the panels to be created
upon initializing the interface. This should be a DataFrame containing
a \code{Name} field specifying the identity of the panel, e.g.,
\code{"Reduced dimension plot 1"}, \code{"Row statistics table 2"}.
The trailing number should not be greater than the number of
maximum plots of that type. The \code{Width} field may also be specified
describing the width of the panel from 2 to 12 (values will be coerced
inside this range).

The \code{tour} argument needs to be provided in a form compatible with the
format expected by the \code{rintrojs} package, where the variables 
\code{element} and \code{intro} have to be present. See more information at
\url{https://github.com/carlganz/rintrojs#usage} regarding the values that 
the \code{element} can assume for proper displaying.
}
\examples{
library(scRNAseq)
data(allen)
class(allen)

# Example data ----

library(scater)
sce <- as(allen, "SingleCellExperiment")
counts(sce) <- assay(sce, "tophat_counts")
sce <- normalize(sce)

sce <- runPCA(sce)
sce <- runTSNE(sce)
rowData(sce)$ave_count <- rowMeans(counts(sce))
rowData(sce)$n_cells <- rowSums(counts(sce)>0)
sce

# launch the app itself ----

app <- iSEE(sce)
if (interactive()) {
  shiny::runApp(app, port = 1234)
}
}
