% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults2.R
\docType{data}
\name{iSEEOptions}
\alias{iSEEOptions}
\title{Default panel options}
\format{An object of class \code{list} of length 4.}
\usage{
iSEEOptions
}
\description{
Set global options using \code{iSEEOptions$set()}, so that all panels will use these default values during initialization.
}
\details{
See \code{str(knitr::iSEEOptions$get())} for a list of default panel options.

Note that \code{iSEEOptions$restore()} can be used to reset the global options to the package default values.
}
\examples{
iSEEOptions$get('downsample'); iSEEOptions$get('selected.color')
}
\keyword{datasets}
