% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{iSEEslots}
\alias{iSEEslots}
\title{iSEE Panel Slot Names}
\description{
Access the name of slots in \pkg{iSEE} panel classes.
}
\details{
The \code{iSEEslots} object exports the name of slots for \pkg{iSEE} panel classes, and should only be used by developers of packages that extend \pkg{iSEE} functionality.
The name of slots should be accessed using the \code{$} operator and the identifier of the slot name, e.g. `iSEEslots$packageVersion`.

Developers of new panels that extend \pkg{iSEE} functionality should add any new slot name to the \code{iSEEslots} object using a new slot identifier, e.g. \code{iSEEslots$myNewSlotIdentifier <- "MyNewSlotName"}.
}
\section{Slot names for the virtual class `Panel`}{

\describe{
\item{\code{packageVersion}}{Named list of package versions active when the panel object was generated or last updated.}
\item{\code{organizationId}}{Integer scalar specifying the width of the panel;
unique across panels of the same concrete class.}
\item{\code{organizationWidth}}{Width of the panel, in Shiny's grid layout units (minimum: 2, maximum 12).}
\item{\code{organizationHeight}}{Height of the panel, in pixels (minimum: 400, maximum 1000).}
}
}

\author{
Kevin Rue-Albrecht
}
