% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_Panel.R
\name{Panel-class}
\alias{Panel-class}
\alias{initialize,Panel-method}
\alias{[[,Panel-method}
\alias{[[<-,Panel-method}
\alias{.refineParameters,Panel-method}
\alias{.cacheCommonInfo,Panel-method}
\alias{.createObservers,Panel-method}
\alias{.hideInterface,Panel-method}
\alias{.multiSelectionRestricted,Panel-method}
\alias{.multiSelectionDimension,Panel-method}
\title{The Panel virtual class}
\description{
The Panel is a virtual base class for all \pkg{iSEE} panels.
It provides slots and methods to control the height and width of each panel,
as well as to control the choice of transmitting panels from which to receive a selection.
}
\section{Slot overview}{

The following slots are relevant to panel organization:
\itemize{
\item \code{PanelId}, an integer scalar specifying the identifier for the panel.
This should be unique across panels of the same concrete class.
\item \code{PanelWidth}, an integer scalar specifying the width of the panel.
Bootstrap coordinates are used so this value should lie between 1 and 12; defaults to 4.
\item \code{PanelHeight}, an integer scalar specifying the height of the panel in pixels.
This is expected to lie between 400 and 1000; defaults to 500.
}

The following slots are relevant to receiving a selection of points:
\itemize{
\item \code{SelectBoxOpen}, a logical scalar indicating whether the selection parameter box should be open.
Defaults to \code{FALSE}.
\item \code{SelectRowSource}, a string specifying the name of the panel from which to receive a selection along the rows (e.g., \code{"rowDataPlot1"}).
Defaults to \code{"---"}.
\item \code{SelectRowType}, a string specifying which row-based selections should be used.
Takes one of \code{"Active"}, only the active selection;
\code{"Union"}, the union of active and saved selections;
and \code{"Saved"}, one of the saved selections.
Defaults to \code{"Active"}.
\item \code{SelectRowSaved}, an integer scalar specifying the index of the saved selection to use when \code{SelectRowType="Saved"}.
Defaults to 0.
\item \code{"SelectColSource"}, a string specifying the name of the panel from which to receive a selection along the columns (e.g., \code{"colDataPlot1"}).
Defaults to \code{"---"}.
\item \code{SelectColType}, a string specifying which column-based selections should be used.
Takes one of \code{"Active"}, only the active selection;
\code{"Union"}, the union of active and saved selections;
and \code{"Saved"}, one of the saved selections.
Defaults to \code{"Active"}.
\item \code{SelectColSaved}, an integer scalar specifying the index of the saved selection to use when \code{SelectColType="Saved"}.
Defaults to 0.
}

The following slots are relevant to transmitting a selection of points:
\itemize{
\item \code{MultiSelectHistory}, a list of arbitrary elements that contain parameters for saved selections.
Each element of this list corresponds to one saved selection.
Defaults to an empty list.
}
}

\section{Getting and setting slots}{

In all of the following code chunks, \code{x} is an instance of a Panel,
and \code{i} is a string containing the slot name:
\itemize{
\item \code{x[[i]]} returns the value of a slot named \code{i}.
\item \code{x[[i]] <- value} modifies \code{x} so that the value in slot \code{i} is replaced with \code{value}.
}
}

\section{Contract description}{

The Panel class offers interface elements to modify all of its slots.
Observers are also created to trigger rerendering upon changes to a slot or when a selection in a transmitting panel changes.

Subclasses are expected to implement methods for, at least:
\itemize{
\item \code{\link{.renderOutput}}
\item \code{\link{.defineOutput}}
\item \code{\link{.getEncodedName}}
\item \code{\link{.getFullName}}
\item \code{\link{.multiSelectionToSave}}
}
Subclasses that transmit selections should also implement specialized methods for selection-related parameters listed below.
}

\section{Supported methods}{

In the following code snippets, \code{x} is an instance of a \linkS4class{ColumnDotPlot} class.
Refer to the documentation for each method for more details on the remaining arguments.

For setting up data values:
\itemize{
\item \code{\link{.refineParameters}(x, se)} is a no-op, returning \code{x} without modification.
\item \code{\link{.cacheCommonInfo}(x, se)} is a no-op, returning \code{se} without modification.
}

For defining the interface:
\itemize{
\item \code{\link{.hideInterface}(x, field)} will always return \code{FALSE}.
}

For controlling selections:
\itemize{
\item \code{\link{.multiSelectionRestricted}(x)} will always return \code{TRUE}.
\item \code{\link{.multiSelectionDimension}(x)} will always return \code{"none"}.
\item \code{\link{.multiSelectionHasActive}(x)} will always return \code{FALSE}.
}
}

\seealso{
\linkS4class{DotPlot} and \linkS4class{Table}, for examples of direct subclasses.
}
\author{
Aaron Lun
}
