% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plot-generics}
\alias{plot-generics}
\alias{.generateDotPlotData}
\alias{.generateDotPlot}
\title{Plotting generics}
\description{
A series of generics for controlling how plotting is performed in \linkS4class{DotPlot} panels.
Subclasses can specialize one or more of them to modify their behavior in specific panels.
}
\section{Generating plotting data}{

In \code{.generateDotPlotData(x, envir)}, the following arguments are required:
\itemize{
\item \code{x}, an instance of a \linkS4class{DotPlot} subclass.
\item \code{envir}, the evaluation environment in which the data.frame is to be constructed.
This can be assumed to have \code{se}, the \linkS4class{SummarizedExperiment} object containing the current dataset;
possibly \code{col_selected}, if a multiple column selection is being transmitted to \code{x};
and possibly \code{row_selected}, if a multiple row selection is being transmitted to \code{x}.
}

The method should add a \code{plot.data} variable in \code{envir} containing a data.frame with columns named \code{"X"} and \code{"Y"}, denoting the variables to show on the x- and y-axes respectively.
It should return a list with \code{commands}, a character vector of commands that produces \code{plot.data} when evaluated in \code{envir};
and \code{labels}, a list of strings containing labels for the x-axis (named \code{"X"}), y-axis (\code{"Y"}) and plot (\code{"title"}).

Each row of the \code{plot.data} data.frame should correspond to one row or column in the SummarizedExperiment \code{envir$se} for \linkS4class{RowDotPlot}s and \linkS4class{ColumnDotPlot}s respectively.
Note that, even if only a subset of rows/columns in the SummarizedExperiment are to be shown, there must always be one row in the data.frame per row/column of the SummarizedExperiment, and in the same order.
All other rows of the data.frame should be filled in with \code{NA}s rather than omitted entirely.
This is necessary for correct interactions with later methods that add other variables to \code{plot.data}.

Any internal variables that are generated by the commands should be prefixed with \code{.} to avoid potential clashes with reserved variable names in the rest of the application.
}

\section{Generating the ggplot object}{

In \code{.generateDotPlot(x, labels, envir)}, the following arguments are required:
\itemize{
\item \code{x}, an instance of a \linkS4class{DotPlot} subclass.
\item \code{labels}, a list of labels corresponding to the columns of \code{plot.data}.
\item \code{envir}, the evaluation environment in which the \link{ggplot} object is to be constructed.
This can be assumed to have \code{plot.data}, a data.frame of plotting data.
}

The method should return a list with \code{plot}, a \link{ggplot} object;
and \code{commands}, a character vector of commands to produce that object when evaluated inside \code{envir}.
This plot will subsequently be the rendered output in \code{\link{.renderOutput}}.

The \code{plot.data} has at least the fields \code{"X"} and \code{"Y"} from \code{\link{.generateDotPlotData}}.
Depending on the parameters of \code{x}, it may also have the following columns:
\itemize{
\item \code{"ColorBy"}, the values of the covariate to ue to color each point. 
\item \code{"ShapeBy"}, the values of the covariate to use for shaping each point.
This is guaranteed to be categorical.
\item \code{"SizeBy"}, the values of the covariate to use for sizing each point.
This is guaranteed to be continuous.
\item \code{"FacetRow"}, the values of the covariate to use to create row facets.
This is guaranteed to be categorical.
\item \code{"FacetColumn"}, the values of the covariate to use to create column facets.
This is guaranteed to be categorical.
\item \code{"SelectBy"}, a logical field indicating whether the point was included in a multiple selection
(i.e., transmitted from another plot with \code{x} as the receiver).
Note that if \code{SelectEffect="Restrict"}, \code{plot.data} will already have been subsetted 
to only retain \code{TRUE} values of this field.
}

\code{envir} may also contain the following variables:
\itemize{
\item \code{plot.data.all}, present when a multiple selection is transmitted to \code{x}
and \code{SelectEffect="Restrict"}.
This is a data.frame that contains all points prior to subsetting and is useful for defining the boundaries of the plot 
such that they do not change when the transmitted multiple selection changes. 
\item \code{plot.data.pre}, present when downsampling is turned on.
This is a data.frame that contains all points prior to downsampling (but after subsetting, if that was performed)
and is again mainly used to fix the boundaries of the plot.
}

Developers may wish to use the \code{\link{.addMultiSelectionPlotCommands}} utility to draw brushes and lassos of \code{x}.
Note that this refers to the brushes and lassos made on \code{x} itself, not those transmitted from another panel to \code{x}.
It would be very unwise for methods to alter the x-axis, y-axis or faceting values in \code{plot.data}.
This will lead to uninuitive discrepancies between apparent visual selections for a brush/lasso
and the actual multiple selection that is evaluated by downstream functions like \code{\link{.processMultiSelections}}.
}

