% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_RedDimPlot.R
\docType{methods}
\name{RedDimPlot}
\alias{RedDimPlot}
\alias{RedDimPlot-class}
\alias{.defineParamInterface,RedDimPlot-method}
\alias{.createParamObservers,RedDimPlot-method}
\title{The reduced dimension plot panel}
\description{
Plots reduced dimensions. What more do I have to say?
}
\section{Constructor}{

\code{RedDimPlot()} creates an instance of a RedDimPlot class.
}

\section{Panel parameters}{

\code{\link{.defineParamInterface}} will create parameter elements for choosing the reduced dimensions to plot.
More details to be added.
}

\examples{
#################
# For end-users #
#################

x <- RedDimPlot()
x[["Type"]]
x[["Type"]] <- "TSNE"

##################
# For developers #
##################

library(scater)
sce <- mockSCE()
sce <- logNormCounts(sce)

# Spits out a NULL and a warning if no reducedDims are available.
sce0 <- .cacheCommonInfo(x, sce)
.refineParameters(x, sce0)

# Replaces the default with something sensible.
sce <- runPCA(sce)
sce0 <- .cacheCommonInfo(x, sce)
.refineParameters(x, sce0)

}
\author{
Aaron Lun
}
