% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reactive.R
\name{.retrieveOutput}
\alias{.retrieveOutput}
\alias{.requestUpdate}
\alias{.requestCleanUpdate}
\alias{.requestActiveSelectionUpdate}
\title{Reactive manipulations for Panel output}
\usage{
.retrieveOutput(panel_name, se, pObjects, rObjects)

.requestUpdate(panel_name, rObjects)

.requestCleanUpdate(panel_name, pObjects, rObjects)

.requestActiveSelectionUpdate(
  panel_name,
  session,
  pObjects,
  rObjects,
  update_output = TRUE
)
}
\arguments{
\item{panel_name}{String containing the panel name.}

\item{se}{A \linkS4class{SummarizedExperiment} object containing the current dataset.}

\item{pObjects}{An environment containing global parameters generated in the \code{\link{iSEE}} app.}

\item{rObjects}{A reactive list of values generated in the \code{\link{iSEE}} app.}

\item{session}{The Shiny session object from the server function.}

\item{update_output}{A logical scalar indicating whether to call \code{.requestUpdate} as well.}
}
\value{
\code{.retrieveOutput} will return the output of running \code{\link{.generateOutput}} for the current panel.

\code{.requestUpdate} will modify \code{rObjects} to request a re-rendering of the specified panel.
\code{.requestCleanUpdate} will also remove all active/saved selections in the chosen panel.

\code{.requestActiveSelectionUpdate} will modify \code{rObjects} to indicate that the active multiple selection for \code{panel_name} has changed.
If \code{update_output=TRUE}, it will also call request a re-rendering of the panel.

All \code{.request*} functions will invisibly return \code{NULL}.
}
\description{
Respond to or request a re-rendering of the \linkS4class{Panel} output via reactive variables.
}
\details{
\code{.retrieveOutput} should be used in the expression for rendering output, e.g., in \code{\link{.renderOutput}}.
This takes care of a number of house-keeping tasks required to satisfy \code{\link{.renderOutput}}'s requirements
with respect to updating various fields in \code{pObjects}.
It also improves efficiency by retrieving cached outputs that were used elsewhere in the app.

\code{.requestUpdate} should be used in various observers to request a re-rendering of the panel,
usually in response to user-driven parameter changes in \code{\link{.createObservers}}.

\code{.requestCleanUpdate} is used for changes to protected parameters that invalidate existing multiple selections,
e.g., if the coordinates change in a \linkS4class{DotPlot}, existing brushes and lassos are usually not applicable.
}
\seealso{
\code{\link{.createProtectedParameterObservers}}, for examples where the update-requesting functions are used.
}
\author{
Aaron Lun
}
