% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_RowDataPlot.R
\docType{methods}
\name{RowDataPlot}
\alias{RowDataPlot}
\alias{RowDataPlot-class}
\alias{.defineParamInterface,RowDataPlot-method}
\alias{.createParamObservers,RowDataPlot-method}
\title{The row data plot panel}
\description{
Plots row data values. What more do I have to say?
}
\section{Constructor}{

\code{RowDataPlot()} creates an instance of a RowDataPlot class.
}

\section{Panel parameters}{

\code{\link{.defineParamInterface}} will create parameter elements for choosing the reduced dimensions to plot.
More details to be added.
}

\examples{
#################
# For end-users #
#################

x <- RowDataPlot()
x[["XAxis"]]
x[["XAxis"]] <- "Row data"

##################
# For developers #
##################

library(scater)
sce <- mockSCE()
sce <- logNormCounts(sce)

# Spits out a NULL and a warning if is nothing to plot.
sce0 <- .cacheCommonInfo(x, sce)
.refineParameters(x, sce0)

# Replaces the default with something sensible.
rowData(sce)$Stuff <- runif(nrow(sce))
sce0 <- .cacheCommonInfo(x, sce)
.refineParameters(x, sce0)

}
\author{
Aaron Lun
}
