% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_SampAssayPlot.R
\docType{methods}
\name{SampAssayPlot}
\alias{SampAssayPlot}
\alias{SampAssayPlot-class}
\alias{.defineParamInterface,SampAssayPlot-method}
\alias{.createParamObservers,SampAssayPlot-method}
\title{The sample assay plot panel}
\description{
Plots sample assay values. What more do I have to say?
}
\section{Constructor}{

\code{SampAssayPlot()} creates an instance of a SampAssayPlot class.
}

\section{Panel parameters}{

\code{\link{.defineParamInterface}} will create parameter elements for choosing the reduced dimensions to plot.
More details to be added.
}

\examples{
#################
# For end-users #
#################

x <- SampAssayPlot()
x[["XAxis"]]
x[["Assay"]] <- "logcounts"
x[["XAxisRowData"]] <- "stuff"

##################
# For developers #
##################

library(scater)
sce <- mockSCE()
sce <- logNormCounts(sce)

old_assay_names <- assayNames(sce)
assayNames(sce) <- character(length(old_assay_names))

# Spits out a NULL and a warning if no assays are named.
sce0 <- .cacheCommonInfo(x, sce)
.refineParameters(x, sce0)

# Replaces the default with something sensible.
assayNames(sce) <- old_assay_names
sce0 <- .cacheCommonInfo(x, sce)
.refineParameters(x, sce0)

}
\author{
Aaron Lun
}
