% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{iSEE selection parameters}
\alias{iSEE selection parameters}
\title{Selection parameters}
\description{
Parameters that control the selection of transmitter plots and the effect of selection in a variety of panels.
}
\section{Selection parameters for plots}{

\describe{
\item{\code{SelectBoxOpen}:}{Logical, should the selection parameter box be open upon initialization?
Defaults to \code{FALSE}.}
\item{\code{SelectByPlot}:}{Character, which other plot should be used for point selection in the current plot?
Defaults to \code{"---"}, which means that no plot is used for point selection.}
\item{\code{SelectEffect}:}{Character, what is the effect of receiving point selection information?
Can be \code{"Restrict"}, where only the selected points are shown; \code{"Color"}, where the selected points have a different color;
or \code{"Transparent"}, where all points other than those selected are made transparent.
Defaults to \code{"Transparent"}.}
\item{\code{SelectColor}:}{Character, what color should be used for the selected points when \code{SelectEffect="Color"}?
Defaults to \code{"red"}.}
\item{\code{SelectAlpha}:}{Numeric, what level of transparency should be used for the unselected points when \code{SelectEffect="Transparent"}?
This should lie in [0, 1], where 0 is fully transparent and 1 is fully opaque.
Defaults to 0.1.}
}
For row-based plots, each point represents a feature, while for column-based plots and heatmaps, each point represents a sample.

Selection information is recorded in the form of brush or lasso objects, which can also be specified:
\describe{
\item{\code{BrushData}:}{A list of shiny brush objects, where each brush object is itself a list - see \code{\link{brushedPoints}} for more details.
The list should be of length equal to the number of plots.}
\item{\code{LassoData}:}{A list of lasso objects, where each lasso object is itself a list - see \code{\link{lassoPoints}} for more details.
The list should be of length equal to the number of plots.}
}
Users should not construct these objects by hand, but instead copy-and-paste the reported code generated by \code{\link{iSEE}}'s code tracker.

It is also possible to store multiple alternative selections for point-based plots through the \code{MultiSelectHistory} field.
This should contain a list of lists of brush or lasso objects containing saved alternative selections for each plot.
The outer list should be of length equal to the number of plots.
Again, users are advised to copy-and-paste reported code rather than writing these objects by hand.

If the transmitting plot has multiple selections, the current panel can choose between them using:
\describe{
\item{\code{SelectMultiType}:}{Character, containing \code{"Active"}, which uses the active selection on the transmitting panel;
\code{"Union"}, which uses the union of the active selection and all saved selections for the transmitting panel;
or \code{"Saved"}, which uses a single saved selection from the transmitting panel.
Defaults to \code{"Active"}.}
\item{\code{SelectMultiSaved}:}{Integer, specifying the saved selection to use from the transmitting panel when \code{SelectMultiType} is set to \code{"Saved"}.
Defaults to 0, i.e., no saved selection is used.}
}
}

\section{Inter-plot transmission rules}{

Point selection cannot occur between row-based and column-based plots.
This is because each point in a row-based plot is a feature, while each point represents a sample in the other plots.
Thus, point selection can only occur between plots of the same point type.
The only exception to this rule is the custom panels, which can receive transmissions from both row- and column-based plots -- see \code{\link{customDataPlotDefaults}}.
This is because they can generate arbitrary plots and are not truly row- or column-based.
}

\section{Selection parameters for tables}{

For row or column statistics tables, the following options apply:
\describe{
\item{\code{SelectBoxOpen}:}{Logical, should the point selection parameter box be open upon initialization?
Defaults to \code{FALSE}.}
\item{\code{SelectByPlot}:}{Character, which other plot should be used to select features in the current table?
Defaults to \code{"---"}, which means that no plot is used for point selection.}
\item{\code{SelectMultiType}:}{Character, containing \code{"Active"}, which uses the active selection on the transmitting panel;
\code{"Union"}, which uses the union of the active selection and all saved selections for the transmitting panel;
or \code{"Saved"}, which uses a single saved selection from the transmitting panel.
}
\item{\code{SelectMultiSaved}:}{Integer, specifying the saved selection to use from the transmitting panel when \code{SelectMultiType} is set to \code{"Saved"}.}
}

Only row-based plots (i.e., row data and sample assay plots) can be used for selecting points to supply to row statistics tables, for the same reasons described above.
Similarly, only column-based plots can be used to select plots to transmit to column statistics tables.
}

\examples{
example(SingleCellExperiment, echo=FALSE) # mock up 'sce'.
redDimPlotDefaults(sce, n=1)
rowStatTableDefaults(sce, n=1)
}
\seealso{
\code{\link{redDimPlotDefaults}},
\code{\link{featAssayPlotDefaults}},
\code{\link{colDataPlotDefaults}},
\code{\link{rowDataPlotDefaults}},
\code{\link{sampAssayPlotDefaults}},
\code{\link{heatMapPlotDefaults}},
\code{\link{rowStatTableDefaults}}
}
\author{
Aaron Lun, Kevin Rue-Albrecht
}
