% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetPointsByGrid.R
\name{subsetPointsByGrid}
\alias{subsetPointsByGrid}
\title{Subset points for faster plotting}
\usage{
subsetPointsByGrid(X, Y, resolution = 200)
}
\arguments{
\item{X}{A numeric vector of x-coordinates for all points.}

\item{Y}{A numeric vector of y-coordinates for all points, of the same length as \code{X}.}

\item{resolution}{A positive integer specifying the number of bins on each axis of the grid.}
}
\value{
A logical vector indicating which points should be retained.
}
\description{
Subset points using a grid-based system, to avoid unnecessary rendering when plotting.
}
\details{
This function will define a grid of the specified resolution across the plot.
Each point is allocated to a grid location (i.e., pair of bins on the x- and y-axes).
If multiple points are allocated to a given location, only the last/right-most point is retained. 
This mimics the fact that plotting will overwrite earlier points with later points.
In this manner, we can avoid unnecessary rendering of earlier points that would not show up anyway.

Note that the \code{resolution} should be a positive integer less than the square root of the maximum integer size,
and will be coerced into the valid range if necessary.
This enables fast calculation of the grid locations for all points.

For plots where \code{X} and \code{Y} are originally categorical, use the jittered versions as input to this function.
}
\examples{
X <- rnorm(100000)
Y <- X + rnorm(100000)

summary(subsetPointsByGrid(X, Y, resolution=100))

summary(subsetPointsByGrid(X, Y, resolution=200))

summary(subsetPointsByGrid(X, Y, resolution=1000))

}
\author{
Aaron Lun
}
