% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentColorMap.R
\name{synchronizeAssays}
\alias{synchronizeAssays}
\title{Synchronise assay color maps to match those in a SummarizedExperiment}
\usage{
synchronizeAssays(ecm, se)
}
\arguments{
\item{ecm}{An \linkS4class{ExperimentColorMap}.}

\item{se}{A \linkS4class{SingleCellExperiment}.}
}
\value{
An \linkS4class{ExperimentColorMap} with color maps in the
\code{assay} slot synchronised to match the position of the corresponding
assay in the \linkS4class{SingleCellExperiment}.
}
\description{
Synchronise assay color maps to match those in a SummarizedExperiment
}
\examples{

# Example ExperimentColorMap ----

count_colors <- function(n){
  c("black","brown","red","orange","yellow")
}
fpkm_colors <- viridis::inferno

ecm <- ExperimentColorMap(
    assays = list(
        counts = count_colors,
        tophat_counts = count_colors,
        cufflinks_fpkm = fpkm_colors,
        rsem_counts = count_colors,
        orphan = count_colors,
        orphan2 = count_colors,
        count_colors,
        tpm_colors
    )
)

# Example SingleCellExperiment ----

library(scRNAseq)
data(allen)
library(scater)
sce <- as(allen, "SingleCellExperiment")
counts(sce) <- assay(sce, "tophat_counts")
sce <- normalize(sce)
sce <- runPCA(sce)
sce <- runTSNE(sce)

# Example ----

ecm_sync <- synchronizeAssays(ecm, sce)

}
