% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_Table.R
\name{Table-class}
\alias{Table-class}
\alias{initialize,Table-method}
\alias{.createObservers,Table-method}
\alias{.renderOutput,Table-method}
\alias{.defineOutput,Table-method}
\alias{.hideInterface,Table-method}
\alias{.multiSelectionCommands,Table-method}
\alias{.multiSelectionActive,Table-method}
\alias{.generateTableCommands}
\title{The Table class}
\description{
The Table is a virtual class for all panels containing a \code{\link{datatable}} widget from the \pkg{DT} package, where each row \emph{usually} corresponds to a row or column of the \linkS4class{SummarizedExperiment} object.
It provides observers for monitoring table selection, global search and column-specific search.
}
\section{Slot overview}{

The following slots control aspects of the \code{DT::datatable} interface:
\itemize{
\item \code{Selected}, a string containing the row name of the currently selected row.
Defaults to \code{NA}, in which case the value should be chosen by the subclass' \code{\link{.refineParameters}} method.
\item \code{Search}, a string containing the regular expression for the global search.
Defaults to \code{""}, i.e., no search.
\item \code{SearchColumns}, a character vector where each entry contains the search string for each column.
Defaults to an empty character vector, i.e., no search.
}

In addition, this class inherits all slots from its parent \linkS4class{Panel} class.
}

\section{Contract description}{

The Table will provide interface elements to create the \code{DT::datatble} widget.
It will also provide observers to update slots based on user interactions with the widget. 
Interface elements and observers are also provided for slots in its parent class \linkS4class{Panel}.

Subclasses are expected to implement methods for (at least) \code{\link{.getTableCommands}}.
See below for more details.
}

\section{Supported methods}{

In the following code snippets, \code{x} is an instance of a \linkS4class{Table} class.
Refer to the documentation for each method for more details on the remaining arguments.

For defining the interface:
\itemize{
\item \code{\link{.defineOutput}(x, id)} returns a UI element for a \code{\link[DT]{dataTableOutput}} widget.
}

For defining reactive expressions:
\itemize{
\item \code{\link{.createObservers}(x, se, input, session, pObjects, rObjects)} sets up observers for all of the slots. 
This will also call the equivalent \linkS4class{Panel} method.
\item \code{\link{.renderOutput}(x, se, colormap, output, pObjects, rObjects)} will add a rendered \code{\link{datatable}} object to \code{output}.
This will also call the equivalent \linkS4class{Panel} method to render the panel information testboxes.
}

For controlling selections: 
\itemize{
\item \code{\link{.multiSelectionRestricted}(x)} returns \code{TRUE}.
Transmission of a selection to a Table will manifest as a subsetting of the rows.
\item \code{\link{.multiSelectionActive}(x)} returns a list containing the contents of \code{x[["Search"]]} and \code{x[["ColumnSearch"]]}.
If both contain only empty strings, a \code{NULL} is returned instead.
\item \code{\link{.multiSelectionCommands}(x, index)} returns a character vector of R expressions that - when evaluated - return a character vector of the row names of the table after applying all search filters.
The value of \code{index} is ignored.
\item \code{\link{.singleSelectionValue}(x, pObjects)} returns the name of the row that was last selected in the \code{\link{datatable}} widget.
}

Unless explicitly specialized above, all methods from the parent class \linkS4class{Panel} are also available.
}

\section{Defining the table commands}{

Developers of Table subclasses do not need to explicitly define \code{\link{.generateOutput}} or \code{\link{.renderOutput}} if they instead specialize \code{\link{.getTableCommands}}.
This generic is used to construct the data.frame that is supplied to the \code{\link{datatable}} widget in \code{\link{.renderOutput,Table-method}}.

In \code{.getTableCommands(x, envir)}, the following arguments are required:
\itemize{
\item \code{x}, an instance of a Table subclass.
\item \code{envir}, the evaluation environment in which the data.frame is to be constructed.
This can be assumed to have \code{se}, the \linkS4class{SummarizedExperiment} object containing the current dataset;
possibly \code{col_selected}, if a multiple column selection is being transmitted to \code{x};
and possibly \code{row_selected}, if a multiple row selection is being transmitted to \code{x}.
}

In return, the method should return a character vector of R commands that, when evaluated in \code{envir}, should yield a data.frame named \code{tab}.
This will automatically be used to construct the \code{\link{datatable}} as well as being stored in \code{pObjects$contents}.

It is permissible to modify the state of \code{envir} as a side-effect;
however, any R expressions that do so should be captured in the output commands.
}

\seealso{
\linkS4class{Panel}, for the immediate parent class.
}
\author{
Aaron Lun
}
